% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot_feature_multiGroup.R
\name{boxPlot_feature_multiGroup}
\alias{boxPlot_feature_multiGroup}
\title{boxPlot_feature_multiGroup}
\usage{
boxPlot_feature_multiGroup(
  processed_se,
  feature,
  ref_group,
  test = c("One-way ANOVA", "Kruskal–Wallis test"),
  post_hoc_sig = c("pval", "padj"),
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{feature}{Character. A feature (lipid) for plotting. It must be one of the lipids in the lipid abundance data.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{test}{Character. The method for comparing means. Allowed method include
"One-way ANOVA" and "Kruskal–Wallis test". Default is \code{'One-way ANOVA'}.}

\item{post_hoc_sig}{Character. The p-value to be used for to determine significance, which is only
applicable for the post hoc test after the Kruskal-Wallis test. Must be one
of "padj" (adjusted p-value) or "pval" (p-value). Default is \code{'pval'}.
##One-way ANOVA只能是padj, Kruskal-Wallis test可以是pval or padj}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", and "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a list of a box plot and a table.
\enumerate{
\item static_boxPlot: abundance box plot of the selected feature.
\item table_boxplot: table for plotting.
\item table_stat: statistical table.
}
}
\description{
This function plots an abundance box plot between multiple groups for a user-selected feature (lipid).
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
boxPlot_result <- boxPlot_feature_multiGroup(
    processed_se, feature='PE O- 17:0;0_20:3;0', ref_group='ctrl',
    test='One-way ANOVA', post_hoc_sig='padj', transform='log10')
}
