% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot_feature_twoGroup.R
\name{boxPlot_feature_twoGroup}
\alias{boxPlot_feature_twoGroup}
\title{boxPlot_feature_twoGroup}
\usage{
boxPlot_feature_twoGroup(
  processed_se,
  feature,
  ref_group,
  test = c("t-test", "Wilcoxon test"),
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by
\code{\link{data_process}}.}

\item{feature}{Character. A feature (lipid) for plotting. It must be one of
the lipids in the lipid abundance data.}

\item{ref_group}{Character. Group name of the reference group. It must be
one of the group names in the group information table's group column.}

\item{test}{Character. The method for comparing means. Allowed method include
"t-test" and "Wilcoxon test". Default is \code{'t-test'}}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", and "cube". Select 'none' to skip data
transformation. Default is \code{'log10'}.}
}
\value{
Return a list of a box plot and a table.
\enumerate{
\item static_boxPlot: abundance box plot of the selected feature.
\item table_boxplot: table for plotting.
\item table_stat: statistical table.
}
}
\description{
This function plots an abundance box plot between two groups
for a user-selected feature (lipid).
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
boxPlot_result <- boxPlot_feature_twoGroup(
    processed_se, feature='Cer 38:1;2', ref_group='ctrl', test='t-test',
    transform='log10')
}
