% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_association.R
\name{char_association}
\alias{char_association}
\title{char_association}
\usage{
char_association(deSp_se, char)
}
\arguments{
\item{deSp_se}{The resulting SummarizedExperiment object from the differential
expression analysis function, such as \code{\link{deSp_twoGroup}} and
\code{\link{deSp_multiGroup}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}.}
}
\value{
Return a list with 3 tables, 3 interactive plots, and 3 static plots.
\enumerate{
\item interactive_barPlot & static_barPlot: a bar plot distinguishes significant groups (values) exhibiting a mean fold change greater than 1.
(NOTE: The bar chart is not generate for multiple-group data.)
\item interacitve_lollipop & static_lollipop: a lollipop plot compares all significant groups
within the selected characteristic by log2(fold change) for two-group data and -log10(p-value) for multiple-group data.
\item interactive_wordCloud & static_wordCloud: a word cloud visualizes the frequency of each group (value) associated with the chosen characteristic.
\item table_barPlot: table for plotting bar plots.
\item table_lollipop: table for plotting lollipop plots.
\item table_wordCloud: table for plotting word cloud.
}
}
\description{
This function shows the significant lipid species based on different
lipid characteristics and visualizes the difference between control and experimental groups.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10' )
deSp_se <- deSp_twoGroup(processed_se, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
result <- char_association(deSp_se, char='class')
}
