% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clin_Cor_heatmap.R
\name{Clin_Cor_heatmap}
\alias{Clin_Cor_heatmap}
\title{Clin_Cor_heatmap}
\usage{
Clin_Cor_heatmap(
  exp_data,
  condition_table,
  test = "pearson",
  adjust_p_method = "BH",
  sig_stat = "p.adj",
  sig_pvalue = 0.05,
  sig_cor_coef = 0.3,
  heatmap_col = "statistic",
  distfun = "spearman",
  hclustfun = "average"
)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class etc.) and their expression of each sample. NAs are allowed. The first column must be 'feature' of gene/lipid names.}

\item{condition_table}{A data frame. The condition table encompasses sample names and clinical conditions (disease status, gene dependence score, etc.)}

\item{test}{A character string indicating which correlation coefficient is to be used for the test. One of \bold{"pearson"}, \bold{"kendall"}, or \bold{"spearman"}, can be abbreviated. (default: "pearson")}

\item{adjust_p_method}{Correction method, a character string. One of \bold{"holm"}, \bold{"hochberg"}, \bold{"hommel"},\bold{"bonferroni"}, \bold{"BH"}, \bold{"BY"}, \bold{"fdr"}, \bold{"none"}, can be abbreviated. (default: "BH")}

\item{sig_stat}{A character string indicating which pvalue is to be used for the statistically significant. One of \bold{"p.adj"} or \bold{"p"}. (default: "p.adj")}

\item{sig_pvalue}{Numeric. Significant level. (default: 1)}

\item{sig_cor_coef}{Numeric. Significance of the correlation coefficient. (default: 0)}

\item{heatmap_col}{A character string indicating the result for printing heatmap. Allow method are \bold{'correlation coefficient'} and \bold{'statistics'}. (default: statistics)}

\item{distfun}{A character string of the distance measure indicating which correlation coefficient (or covariance) is to be computed. Allowed methods include \bold{"pearson"}, \bold{"kendall"}, \bold{"spearman"}.(default: "spearman")}

\item{hclustfun}{A character string of the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of \bold{"ward.D"}, \bold{"ward.D2"}, \bold{"single"}, \bold{"complete"}, \bold{"average"}(= UPGMA), \bold{"mcquitty"} (= WPGMA), \bold{"median"} (= WPGMC) or \bold{"centroid"} (= UPGMC). (default: "average")}
}
\value{
Return a list with 2 data frames, 1plot, and 1 matrix.
\enumerate{
\item Clin_Cor_table_all: data frame, correlation table of clinical features and lipid species.
\item Clin_Cor_table_sig: data frame, correlation table of significant clinical features and lipid species.
\item Clin_Cor_table_plot: heatmap of the correlation coefficient between clinical features (e.g. genes) and lipid species.
\item Clin_Cor_reorder_mat: clinical features and lipid species correlation reorder matrix.
}
}
\description{
The correlation coefficient provides a summary view of the existing relationship between clinical features and lipid species, how strong that relationship is, and whether the relationship is positive or negative. Three types of correlations, Pearson, Spearman, and Kendall, are provided. The cut-offs for the correlation coefficient and the p-value can be decided by users.

This function returns the heatmap after users input the cut-offs and choose a value for clustering/methods for clustering. Users can use either correlation coefficient between clinical features (e.g. genes) and lipid species or choose their statistic instead.
}
\examples{
data("corr_exp_data")
data("corr_condition_table")
exp_data <- corr_exp_data
condition_table <- corr_condition_table
exp_transform <- data_process(exp_data, exclude_var_missing=TRUE,
                              missing_pct_limit=50, replace_zero=TRUE,
                              zero2what='min', xmin=0.5, replace_NA=TRUE,
                              NA2what='min', ymin=0.5, pct_transform=TRUE,
                              data_transform=TRUE, trans_type='log',
                              centering=FALSE,  scaling=FALSE)
Clin_Cor_heatmap(exp_transform, condition_table, test = 'pearson',
                 adjust_p_method = 'BH', sig_stat = 'p.adj',
                 sig_pvalue = 1, sig_cor_coef = 0, heatmap_col='statistic',
                 distfun='spearman', hclustfun='average')
}
