% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sp2char.R
\name{convert_sp2char}
\alias{convert_sp2char}
\title{convert_sp2char}
\usage{
convert_sp2char(processed_se, transform = c("none", "log10", "square", "cube"))
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", and "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object.
}
\description{
This function sums up lipid species abundance according to lipid characteristics,
which converts the lipid species abundance table to the lipid characteristic abundance table.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_data <- convert_sp2char(processed_se, transform='log10')
}
