% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_sp2ratio.R
\name{convert_sp2ratio}
\alias{convert_sp2ratio}
\title{convert_sp2ratio}
\usage{
convert_sp2ratio(processed_se, transform = c("none", "log2"))
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{transform}{Character. Method for ratio-based abundance transformation.
Allowed methods include "none" and "log2". Select 'none' to skip data transformation. Default is \code{'log2'}.}
}
\value{
Return a SummarizedExperiment object.
}
\description{
This function is designed to convert species abundance into
specific ratio-based abundance.
}
\examples{
data("de_data_twoGroup")
processed_se_twoGroup <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
ratioAbund_twoGroup <- convert_sp2ratio(
    processed_se_twoGroup, transform='log2')

data("se_multiGroup")
processed_se_multiGroup <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
ratioAbund_multiGroup <- convert_sp2ratio(
    processed_se_multiGroup, transform='log2')
}
