\name{corr_adjusted_table}
\alias{corr_adjusted_table}
\docType{data}
\title{
corr_adjusted_table
}
\description{
A data frame with additional variables to adjust confounding effects for correlation analysis workflow.
}
\usage{data("corr_adjusted_table")}
\format{
  A data frame with 129 observations on the following 6 variables.
  \describe{
    \item{\code{sample_name}}{a character vector}
    \item{\code{Age}}{a numeric vector}
    \item{\code{Sex}}{a numeric vector}
    \item{\code{Smoking}}{a numeric vector}
    \item{\code{BMI}}{a numeric vector}
    \item{\code{FEV1}}{a numeric vector}
  }
}
\examples{
data(corr_adjusted_table)
}
\keyword{datasets}
