\name{corr_condition_table}
\alias{corr_condition_table}
\docType{data}
\title{
corr_condition_table
}
\description{
A data frame of sample names and clinical conditions for correlation analysis workflow.
}
\usage{data("corr_condition_table")}
\format{
  A data frame with 129 observations on the following 4 variables.
  \describe{
    \item{\code{sample_name}}{a character vector}
    \item{\code{FEV1_FVC}}{a numeric vector}
    \item{\code{Emphysema}}{a numeric vector}
    \item{\code{Exacerbations}}{a numeric vector}
  }
}
\examples{
data(corr_condition_table)
}
\keyword{datasets}
