% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{corr_heatmap}
\usage{
corr_heatmap(
  exp_data,
  corr_method = "pearson",
  distfun = "maximum",
  hclustfun = "average",
  plotly = TRUE,
  type = "sample"
)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features
(molecule, lipid class, etc.) and their expression of each sample. NAs are
not allowed. The name of the first column must be "feature" (lipid species).}

\item{corr_method}{A character string indicating which correlation
coefficient is to be computed. One of "pearson" or
"spearman". (default: "pearson")}

\item{distfun}{A character string of the distance measure indicating which
correlation coefficient (or covariance) is to be computed. Allowed methods
include "pearson", "spearman", "kendall", euclidean", "maximum",
"manhattan", "canberra", "binary", "minkowski". (default: "maximum")}

\item{hclustfun}{A character string of the agglomeration method to be used.
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2",
"single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC). (default: "average")}

\item{plotly}{Logical value. If TRUE, return the resulting plots dynamically.}

\item{type}{It should be 'sample' or 'lipid.' 'sample' outputs the
correlation results of samples, and 'lipid' outputs output the correlation
results of lipid species.}
}
\value{
Return 1 plot and a list of 3 matrices.
\enumerate{
\item corr_coef: a matrix of correlation coefficients between samples
or lipids.
\item corr_p: a matrix of correlation p-value between samples or lipids.
\item heatmap: a heatmap plot by samples or lipid species.
\item reorder_data: a reorder sample or lipids correlation coefficients
matrix.
}
}
\description{
The correlation heatmap illustrates the correlation between
lipid species or samples and depicts the patterns in each group.
The correlation can be calculated by Pearson or Spearman.
\enumerate{
\item Compute and cluster the correlation coefficient depending on
user-defined method and distance.
\item Print two correlation heatmaps to illustrate the correlation between
lipid species or samples.
}
}
\examples{
data("profiling_exp_data")
exp_data <- profiling_exp_data[1:50, ]
exp_transform <- data_process(exp_data, exclude_var_missing=TRUE,
                              missing_pct_limit=50,
                              replace_zero=TRUE, zero2what='min',
                              xmin=0.5, replace_NA=TRUE,
                              NA2what='min', ymin=0.5, pct_transform=TRUE,
                              data_transform=TRUE, trans_type='log',
                              centering=FALSE, scaling=FALSE)
corr_heatmap(exp_transform, corr_method="pearson", distfun="maximum",
             hclustfun="average", plotly=TRUE, type='sample')
}
