% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_lr_heatmap.R
\name{corr_lr_heatmap}
\alias{corr_lr_heatmap}
\title{corr_lr_heatmap}
\usage{
corr_lr_heatmap(
  processed_se,
  char = NULL,
  condition_col,
  adjusted_col,
  side_color_char,
  significant = c("padj", "pval"),
  p_cutoff = 0.05,
  adjust_p_method = c("holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
    "none"),
  distfun = c("pearson", "spearman", "kendall"),
  hclustfun = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  heatmap_col = c("beta_coef", "t_statistic"),
  transform = c("none", "log10", "square", "cube"),
  type = c("Sp", "Char")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}.}

\item{condition_col}{Character.The column names used to extract the condition
table from the group information table, including clinical conditions such as
disease status or gene dependency scores.}

\item{adjusted_col}{Character. The column names used to extract the adjusted table from
the group information table, including additional variables to be incorporated
into the algorithm for adjusting confounding effects.}

\item{side_color_char}{Character. A lipid characteristic used for plotting
the side color of heatmap. It must be selected from the common list  returned
by \code{\link{list_lipid_char}}.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. Significant level. Default is \code{1}.}

\item{adjust_p_method}{Character. The correction method of p-value. Allowed
methods include "holm", "hochberg", "hommel", "bonferroni", "BH", "BY", "fdr",
and "none". Default is \code{'BH'}.}

\item{distfun}{Character. The distance measure for computing correlation
coefficient (or covariance). Allowed methods include "pearson", "kendall",
"spearman". Default is \code{'spearman'}.}

\item{hclustfun}{Character. The agglomeration method. This should be
(an unambiguous abbreviation of) one of "ward.D", "ward.D2",
"single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC). Default is \code{'centroid'}.}

\item{heatmap_col}{Character. The value for clustering. Allow method are
"beta_coef" and "t_statistic". Default is \code{'t_statistic'}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}

\item{type}{Character. Specifies the correlation type: 'Sp' for lipid species
correlation and 'Char' for lipid characteristic correlation.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function uses multiple explanatory variables to predict the
outcome of a continuous response variable using linear regression. It enables
researchers to estimate associations between lipid levels and clinical features.
}
\examples{
data("corr_data")
processed_se <- data_process(
    corr_data, exclude_missing=TRUE, exclude_missing_pct=70, replace_na_method='min',
    replace_na_method_ref=0.5, normalization='Percentage', transform='log10')
result <- corr_lr_heatmap(processed_se, char=NULL,
    condition_col=c("FEV1_FVC", "Emphysema", "Exacerbations"),
    adjusted_col=c("Age", "Sex", "Smoking", "BMI", "FEV1"),
    side_color_char=NULL, significant='pval', p_cutoff=0.05,
    adjust_p_method='BH', distfun='spearman', hclustfun='centroid',
    heatmap_col='t_statistic', transform='log10', type='Sp')
}
