% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_sample_variability.R
\name{cross_sample_variability}
\alias{cross_sample_variability}
\title{cross_sample_variability}
\usage{
cross_sample_variability(se)
}
\arguments{
\item{se}{A SummarizedExperiment object construct by \code{\link{as_summarized_experiment}}.}
}
\value{
Return a list of 3 interactive plots, 3 static plots, and 2 tables.
\enumerate{
\item interactive_lipid_number_barPlot & static_lipid_number_barPlot: histogram of lipid numbers.
\item interactive_lipid_amount_barPlot & static_lipid_amount_barPlot: histogram of the total amount of lipid in each sample.
\item interactive_lipid_distribution & static_lipid_distribution: density plot of the underlying probability distribution of the lipid abundance in each sample.
\item table_total_lipid: table for plotting lipid_number_barPlot & lipid_amount_barPlot.
\item table_lipid_distribution: table for plotting lipid_distribution.
}
}
\description{
This function provides three distribution plots for users to view sample
variability and compare lipids' amount/abundance difference between samples (i.e., patients vs. control).
}
\examples{
data("profiling_data")
result <- cross_sample_variability(profiling_data)
}
