% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{data_process}
\usage{
data_process(
  exp_data,
  exclude_var_missing = T,
  missing_pct_limit = 50,
  replace_zero = T,
  zero2what = "min",
  xmin = 0.5,
  replace_NA = T,
  NA2what = "min",
  ymin = 0.5,
  pct_transform = T,
  data_transform = T,
  trans_type = "log",
  centering = F,
  scaling = F
)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}

\item{exclude_var_missing}{Logical. Remove Lipid with too many missing values. (default: T)}

\item{missing_pct_limit}{An integer indicating the missing values over a certain percentage should be removed. (default: 50)}

\item{replace_zero}{Logical. Replace 0. (default: T)}

\item{zero2what}{A character string indicating the value to replace 0. Value include "mean", "median", "min". (default: min)}

\item{xmin}{A numeric value indicating the min value to replace 0.}

\item{replace_NA}{Logical. If remove_na = TRUE, all NA will be removed. (default: T)}

\item{NA2what}{A character string indicating the value to replace NA. Value include "mean", "median", "min". (default: min)}

\item{ymin}{A numeric value indicating the min value to replace NA.}

\item{pct_transform}{Logical. If \bold{pct_transform = TRUE}, transform lipid value into a percentage. (default:T)}

\item{data_transform}{Logical. If \bold{data_transform = TRUE}, transform exp_data by log10.}

\item{trans_type}{A character string of transformation type.}

\item{centering}{A logical value indicating whether the variables should be shifted to be zero centered. AAlternately, a vector of length equal to the number of columns of x can be supplied. The value is passed to scale. (default: F)}

\item{scaling}{A logical value. If scaling = TRUE, each block is standardized to zero means and unit variances. (default: F).}
}
\value{
Return a list of 1 data frame.
}
\description{
This function conducts data processing according to users' options, including removing features with missing values, missing values imputation, percentage transformation, log10 transformation, etc.
}
\examples{
\donttest{
data("DE_exp_data")
exp_data <- DE_exp_data
data_process(exp_data, exclude_var_missing=T, missing_pct_limit=50, replace_zero=T,
             zero2what='min', xmin=0.5, replace_NA=T, NA2what='min', ymin=0.5,
             pct_transform=T, data_transform=T, trans_type='log', centering=F, scaling=F)
}
}
