% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_char.R
\name{DE_char}
\alias{DE_char}
\title{DE_char}
\usage{
DE_char(
  exp_data_SE,
  data_transform = TRUE,
  paired = FALSE,
  sig_pvalue = 0.05,
  sig_FC = 2,
  insert_ref_group = NULL,
  ref_group = NULL
)
}
\arguments{
\item{exp_data_SE}{A SummarizedExperiment object contains information
about various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
allowed. The row data corresponds to specific lipid features, such as class
and total length. The first  column's name must be "feature" (lipid
species), and NAs are allowed for this data. The column data comprises
sample names, sample labels, group names, and  pair numbers that represent
'the pair' for conducting t-tests or Wilcoxon  tests. NAs are allowed.
\emph{NOTE: The output of Species2Char}.}

\item{data_transform}{Logical. If data_transform = TRUE, transformed
exp_data by log10.}

\item{paired}{Logical. If \bold{paired = TRUE}, data are paired samples.
(default: FALSE)}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{sig_FC}{Numeric. Significance of the fold-change. (default: 2)}

\item{insert_ref_group}{A character string. The name of 'ctrl' after name
conversion.}

\item{ref_group}{A character string. The name of 'exp' after name conversion.}
}
\value{
Return a list with 5 SummarizedExperiment objects.
\enumerate{
\item char_exp_data: the expression of the user-selected lipid
characteristics by samples.
\item char_table_all: statistical analysis for differentially expressed
characters. For example, the result of two-way ANOVA will tell you whether
the interaction effect between totallength and different groups is present.
The post hoc test (t-test) will then calculate significance for each length
of FA and produce a p-value.
\item combined_table: the value calculated by the weighted average of the
expression of the continuous lipid characteristics.
\item combine_result_table: statistics of t.test for control and experiment
groups.
\item bar_table: data that can be further applied for plotting bar plot.
\item char_box: data that can be further applied for plotting box plot.
\item char_table_all_sig: the significant results of statistical analysis
for differentially expressed characters.
}
}
\description{
The \code{\link{DE_char}} function is for lipid characteristics analysis,
which is used to search for deferentially expressed lipid characters based
on the user's choice. In this function, two-way ANOVA will be used to
evaluate the interaction between lipid characteristics (class, total length,
etc.) and group information (control v.s experimental group) on the lipid
expression values or percentages (according to selected characteristics). To
obtain the overview of differentially expressed lipids by user-selected
characteristics, an expression table from \code{\link{Species2Char}} and a
lipid characteristic table are needed. Note that the sample can be paired or
not, depending on \bold{paired=TRUE/FALSE}.
}
\examples{
library(SummarizedExperiment)
data("DE_data")
exp_data_SE <- DE_data
lipid_char_table <- as.data.frame(rowData(exp_data_SE))
char_var <- colnames(lipid_char_table)[-1]
Spe2Char_result <- Species2Char(exp_data_SE, char_var = char_var[4])
exp_transform_non_log <- data_process(Spe2Char_result,
    exclude_var_missing=TRUE, missing_pct_limit=50, replace_zero=TRUE,
    zero2what='min', xmin=0.5, replace_NA=TRUE, NA2what='min', ymin=0.5,
    pct_transform=TRUE, data_transform=FALSE, trans_type='log',
    centering=FALSE, scaling=FALSE)
DE_char(exp_transform_non_log, data_transform = TRUE, paired = FALSE,
    sig_pvalue = 0.05, sig_FC = 2, insert_ref_group=NULL, ref_group=NULL)
}
