% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_char_2.R
\name{DE_char_2}
\alias{DE_char_2}
\title{DE_char_2}
\usage{
DE_char_2(
  exp_data,
  data_transform = T,
  group_info,
  paired = F,
  sig_pvalue = 0.05,
  sig_FC = 2,
  insert_ref_group = NULL,
  ref_group = NULL,
  char_var = "Category"
)
}
\arguments{
\item{exp_data}{A data frame with a list of lipid caracteristics in the first row and sample names as column name with expression as value. \emph{NOTE: The output of Species2Char}.}

\item{data_transform}{Logical. If data_transform = TRUE, transformed exp_data by log10.}

\item{group_info}{A data frame according to Group information table (after convert to ctrl and exp). NAs are allowed.}

\item{paired}{Logical. If \bold{paired = T}, data are paired samples. (default: F)}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{sig_FC}{Numeric. Significance of the fold-change. (default: 2)}

\item{insert_ref_group}{A character string. The name of 'ctrl' after name conversion.}

\item{ref_group}{A character string. The name of 'exp' after name conversion.}

\item{char_var}{A character string of the lipid characteristic selected by users from the column name of \bold{lipid_char_table}.}
}
\value{
Return a list with 4 data frames and 5 figures.
\enumerate{
\item DE_char_exp_data: data frame, the expression of the user-selected lipid characteristics by samples.
\item DE_char_table_all: data frame with statistical analysis for differentially expressed characters. For example, the result of two-way ANOVA will tell you whether the interaction effect between totallength and different groups is present. The post hoc test (t-test) will then calculate significance for each length of FA and produce a p-value.
\item DE_char_combined_table: data frame with the value calculated by the weighted average of the expression of the continuous lipid characteristics. \emph{NOTE: this output is used for DE_char_trendplot}.
\item DE_char_combine_result_table: data frame with statistics of t.test for control and experiment groups.
\item DE_char_barplot: a barplot shows the average of each sample group and whether they are significantly differentially expressed between two groups based on user-selected character.
\item DE_char_barplot_sqrt: a barplot shows the average of each sample group and whether they are significantly differentially expressed between two groups based on user-selected character. \emph{NOTE: the y axis is sqrt-scaled}.
\item DE_char_boxplot: plot with the index value from DE_char_combined_table and DE_char_combine_result_table.
\item DE_char_trendplot: a line chart shows the average of each sample group and whether they are significantly differentially expressed between two groups based on user-selected character.
\item DE_char_trendplot_sqrt: a line shows the average of each sample group and whether they are significantly differentially expressed between two groups based on user-selected character. \emph{NOTE: the y axis is sqrt-scaled}.
}
}
\description{
The \code{\link{DE_sub_char_2}} function is for lipid characteristics analysis, which is used to search for deferentially expressed lipid characters based on the user's choice. In this function, two-way ANOVA will be used to evaluate the interaction between lipid characteristics (class, total length, etc.) and group information (control v.s experimental group) on the lipid expression values or percentages (according to selected characteristics). To obtain the overview of differentially expressed lipids by user-selected characteristics, an expression table from \code{\link{Species2Char}} and a lipid characteristic table are needed. Note that the sample can be paired or not, depending on \bold{paired=T/F}, and this function can further be used for subgroup analysis, \code{\link{DE_sub_char_2}}.
\enumerate{
\item The analysis of the user-selected characteristics.
\item Print a box plot, line plots (raw & sqrt scale), and bar plots (raw & sqrt scale).
}
}
\examples{
\dontrun{
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
exp_data <- DE_exp_data
lipid_char_table <- DE_lipid_char_table
group_info <- DE_group_info
char_var <- colnames(lipid_char_table)[-1]
exp_data_Spe2Char <- Species2Char(exp_data, lipid_char_table = lipid_char_table,
                                  char_var = char_var[4])
exp_transform_non_log <- data_process(exp_data_Spe2Char, exclude_var_missing=T,
                                      missing_pct_limit=50, replace_zero=T, zero2what='min',
                                      xmin=0.5, replace_NA=T, NA2what='min', ymin=0.5,
                                      pct_transform=T, data_transform=F, trans_type='log',
                                      centering=F, scaling=F)
DE_char_2(exp_transform_non_log, data_transform = T, group_info = group_info, paired = F,
          sig_pvalue = 0.05, sig_FC = 2, insert_ref_group=NULL, ref_group=NULL)
}
}
