\name{DE_group_info}
\alias{DE_group_info}
\docType{data}
\title{
DE_group_info
}
\description{
A data frame of one clinical term (or one set of clinical terms) from demographic data for differential expression analysis workflow.
}
\usage{data("DE_group_info")}
\format{
  A data frame with 23 observations on the following 4 variables.
  \describe{
    \item{\code{sample_name}}{a character vector}
    \item{\code{label_name}}{a character vector}
    \item{\code{group}}{a character vector}
    \item{\code{pair}}{a logical vector}
  }
}
\examples{
data(DE_group_info)
}
\keyword{datasets}
