\name{DE_lipid_gene_path}
\alias{DE_lipid_gene_path}
\docType{data}
\title{
DE_lipid_gene_path
}
\description{
A data frame of pathway information.
}
\usage{data("DE_lipid_gene_path")}
\value{
  A data frame with 28148 observations on the following 10 variables.
  \describe{
    \item{\code{key_name}}{a character vector}
    \item{\code{DB}}{a character vector}
    \item{\code{ID_type}}{a character vector}
    \item{\code{ID}}{a character vector}
    \item{\code{KEGG_ID}}{a character vector}
    \item{\code{CHEBI_ID}}{a numeric vector}
    \item{\code{path_id}}{a character vector}
    \item{\code{path_name}}{a character vector}
    \item{\code{gene_id}}{a character vector}
    \item{\code{gene_name}}{a character vector}
  }
}
\examples{
data(DE_lipid_gene_path)
}
\keyword{datasets}
