% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sub_char.R
\name{DE_sub_char}
\alias{DE_sub_char}
\title{DE_sub_char}
\usage{
DE_sub_char(
  exp_data_SE,
  data_transform = TRUE,
  split_var,
  char_var,
  paired = FALSE,
  sig_pvalue = 0.05,
  sig_FC = 0,
  exclude_var_missing = TRUE,
  missing_pct_limit = 50,
  replace_zero = TRUE,
  zero2what = "min",
  xmin = 0.5,
  replace_NA = TRUE,
  NA2what = "min",
  ymin = 0.5,
  pct_transform = TRUE,
  trans_type = "log",
  centering = FALSE,
  scaling = FALSE
)
}
\arguments{
\item{exp_data_SE}{A SummarizedExperiment object contains information about
various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
allowed. The row data  corresponds to specific lipid features, such as class
and total length. The first column's name must be "feature" (lipid species),
and NAs are allowed for this data. The column data comprises sample names,
sample labels, group names, and pair numbers that represent 'the pair' for
conducting t-tests or Wilcoxon tests. NAs are allowed.}

\item{data_transform}{Logical. If data_transform = TRUE, transform exp_data
by log10.}

\item{split_var}{A character string of the second lipid characteristic
selected by users from the column name of the rowData of \bold{exp_data_SE}
 for the subgroup analysis, such as class. \emph{NOTE: This parameter will
 be used to split data before entering the main lipid characteristic '
 char_var'.}}

\item{char_var}{A character string of the first lipid characteristic
selected by users from tthe column name of the rowData of
\bold{exp_data_SE}, such as total length.}

\item{paired}{Logical. If paired = TRUE, data are paired samples.
(default: FALSE)}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{sig_FC}{Numeric. Significance of the fold-change. (default: 0)}

\item{exclude_var_missing}{Logical. Remove Lipid with too many missing
values. (default: TRUE)}

\item{missing_pct_limit}{An integer indicating the missing values over a
certain percentage should be removed. (default: 50)}

\item{replace_zero}{Logical. Replace 0. (default: TRUE)}

\item{zero2what}{A character string indicating the value to replace 0. Value
include "mean", "median", "min". (default: min)}

\item{xmin}{A numeric value indicating the min value to replace 0.}

\item{replace_NA}{Logical. If remove_na = TRUE, all NA will be removed.
(default: TRUE)}

\item{NA2what}{A character string indicating the value to replace NA. Value
include "mean", "median", "min". (default: min)}

\item{ymin}{A numeric value indicating the min value to replace NA.}

\item{pct_transform}{Logical. If \bold{pct_transform = TRUE}, transform
lipid value into a percentage. (default:TRUE)}

\item{trans_type}{Logical. If \bold{data_transform = TRUE}, transform
exp_data by log10.}

\item{centering}{A logical value indicating whether the variables should be
shifted to be zero centered. Alternately, a vector of length equal to the
number of columns of x can be supplied. The value is passed to scale.
(default: FALSE)}

\item{scaling}{A logical value. If scaling = TRUE, each block is
standardized to zero means and unit variances. (default: FALSE).}
}
\value{
Return a list with 4 SummarizedExperiment objects.
\enumerate{
\item result_table1: scaled expression based on two user-selected lipid
characters' grouping
\item result_table2: two-way anova result with a post_hoc_test result.
\item result_table3: scaled expression based on the char_var's weighted mean.
\item result_table4: t.test that based on result_table3.
}
}
\description{
This function conduct lipid Characteristics Analysis.
Lipid species are categorized and summarized into a new lipid expression
table according to two selected lipid characteristics, then conducted
differential expressed analysis.
}
\examples{
data("DE_data")
exp_data_SE <- DE_data
lipid_char_table <- as.data.frame(SummarizedExperiment::rowData(exp_data_SE))
char_var <- colnames(lipid_char_table)[-1]
DE_sub_char(exp_data_SE, data_transform=TRUE, split_var = char_var[2],
char_var = char_var[4], paired = FALSE, sig_pvalue=0.05, sig_FC=2,
exclude_var_missing=TRUE, missing_pct_limit=50, replace_zero=TRUE,
zero2what='min', xmin=0.5, replace_NA=TRUE, NA2what='min', ymin=0.5,
pct_transform=TRUE, trans_type='log', centering=FALSE, scaling=FALSE)
}
