% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sub_char_2.R
\name{DE_sub_char_2}
\alias{DE_sub_char_2}
\title{DE_sub_char_2}
\usage{
DE_sub_char_2(
  exp_data,
  data_transform = TRUE,
  lipid_char_table,
  split_var,
  char_var,
  group_info,
  paired = FALSE,
  sig_pvalue = 0.05,
  sig_FC = 0,
  exclude_var_missing = TRUE,
  missing_pct_limit = 50,
  replace_zero = TRUE,
  zero2what = "min",
  xmin = 0.5,
  replace_NA = TRUE,
  NA2what = "min",
  ymin = 0.5,
  pct_transform = TRUE,
  trans_type = "log",
  centering = FALSE,
  scaling = FALSE
)
}
\arguments{
\item{exp_data}{A data frame includes the expression of lipid features in
each sample. NAs are allowed. First column should be gene/lipid name and
first column name must be 'feature'.}

\item{data_transform}{Logical. If data_transform = TRUE, transform exp_data
by log10.}

\item{lipid_char_table}{A data frame with lipid features, such as class,
total length. NAs are allowed. The name of the first column must be
"feature".}

\item{split_var}{A character string of the second lipid characteristic
selected by users from \bold{lipid_char_table} for the subgroup analysis,
such as class. \emph{NOTE: This parameter will be used to split data before
entering the main lipid characteristic 'char_var'.}}

\item{char_var}{A character string of the first lipid characteristic
selected by users from the column name of \bold{lipid_char_table},
such as total length.}

\item{group_info}{A data frame comprises the name of the sample,
the label of the sample, the group name of the sample, and the pair number
represents 'the pair' for the t-test/Wilcoxon test. NAs are allowed.}

\item{paired}{Logical. If paired = TRUE, data are paired samples.
(default: FALSE)}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{sig_FC}{Numeric. Significance of the fold-change. (default: 0)}

\item{exclude_var_missing}{Logical. Remove Lipid with too many missing
values. (default: TRUE)}

\item{missing_pct_limit}{An integer indicating the missing values over a
certain percentage should be removed. (default: 50)}

\item{replace_zero}{Logical. Replace 0. (default: TRUE)}

\item{zero2what}{A character string indicating the value to replace 0. Value
include "mean", "median", "min". (default: min)}

\item{xmin}{A numeric value indicating the min value to replace 0.}

\item{replace_NA}{Logical. If remove_na = TRUE, all NA will be removed.
(default: TRUE)}

\item{NA2what}{A character string indicating the value to replace NA. Value
include "mean", "median", "min". (default: min)}

\item{ymin}{A numeric value indicating the min value to replace NA.}

\item{pct_transform}{Logical. If \bold{pct_transform = TRUE}, transform
lipid value into a percentage. (default:TRUE)}

\item{trans_type}{Logical. If \bold{data_transform = TRUE}, transform
exp_data by log10.}

\item{centering}{A logical value indicating whether the variables should be
shifted to be zero centered. AAlternately, a vector of length equal to the
number of columns of x can be supplied. The value is passed to scale.
(default: FALSE)}

\item{scaling}{A logical value. If scaling = TRUE, each block is
standardized to zero means and unit variances. (default: FALSE).}
}
\value{
Return a list with 4 data frames.
\enumerate{
\item result_table1: a data frame with scaled expression based on two
user-selected lipid characters' grouping
\item result_table2: a data frame with two-way anova result with a
post_hoc_test result
\item result_table3: a data frame of scaled expression based on the
char_var's weighted mean
\item result_table4: a data frame of t.test that based on result_table3
}
}
\description{
In the \code{\link{DE_sub_char_2}}, lipid Characteristics
Analysis will be conducted. Lipid species are categorized and summarised
into a new lipid expression table according to two selected
lipid characteristics, then conducted differential expressed analysis. \\
Note:
\enumerate{
\item The subgroup analysis is based on the first user-selected
characteristics. So, use \code{\link{DE_char_2}} before using
\code{\link{DE_sub_char_2}},
\item two selected characteristics should be both continuous data and one
categorical data with one continuous data. The cut-offs of differentially
expressed lipids are inputted by users.
}
}
\examples{
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
group_info <- DE_group_info[c(1:5, 11:15), ]
exp_data <- DE_exp_data[1:100, c("feature", group_info$sample_name)]
lipid_char_table <- DE_lipid_char_table[1:100, ]
char_var <- colnames(lipid_char_table)[-1]
DE_sub_char_2(exp_data, data_transform=TRUE,
              lipid_char_table = lipid_char_table,
              split_var = char_var[2], char_var = char_var[6],
              group_info = group_info,
              paired = FALSE, sig_pvalue=0.05, sig_FC=2,
              exclude_var_missing=TRUE, missing_pct_limit=50,
              replace_zero=TRUE, zero2what='min', xmin=0.5,
              replace_NA=TRUE, NA2what='min', ymin=0.5, pct_transform=TRUE,
              trans_type='log', centering=FALSE, scaling=FALSE)
}
