% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_sub_char_plot_2.R
\name{DE_sub_char_plot_2}
\alias{DE_sub_char_plot_2}
\title{DE_sub_char_plot_2}
\usage{
DE_sub_char_plot_2(
  DE_split_char_table_all,
  DE_split_char_index,
  group_info,
  char_var = "Category",
  split_var,
  split_class,
  insert_ref_group = NULL,
  ref_group = NULL
)
}
\arguments{
\item{DE_split_char_table_all}{A data frame. The output of \code{\link{DE_sub_char_2}}.}

\item{DE_split_char_index}{A data frame. The output of \code{\link{DE_sub_char_2}}.}

\item{group_info}{A data frame comprises the name of the sample, the label of the sample, the group name of the sample, and the pair number represents ‘the pair’ for the t-test/Wilcoxon test. NAs are allowed. NAs are allowed.}

\item{char_var}{A character string of the first lipid characteristic selected by users from the column name of \bold{lipid_char_table}, such as total length.}

\item{split_var}{A character string of the second lipid characteristic selected by users from \bold{lipid_char_table} for the subgroup analysis, such as class. \emph{NOTE: This parameter will be used to split data before entering main lipid characteristic 'char_var'.}}

\item{split_class}{A character string selected by users from the second user-selected lipid characteristic (\bold{split_var}) to visualise the specific plot for the selected category of that characteristic. For example, when user set split_var= 'class', and split_class='PC', this indicates the results will be plots of PC.}

\item{insert_ref_group}{A character string. The name of 'ctrl' after name conversion.}

\item{ref_group}{A character string. The name of 'exp' after name conversion.}
}
\value{
Return a list with 5 figures.
\enumerate{
\item a bar plot of \bold{split_class}
\item a line plot of \bold{split_class}
\item a box plot of \bold{split_class}
\item a the bar plot of \bold{split_class} with the sqrt scale
\item a line plot of \bold{split_class} with the sqrt scale
}
}
\description{
This \code{\link{DE_sub_char_plot_2}} is to visualise the result of \code{\link{DE_sub_char_2}}.
\enumerate{
\item The subgroup analysis of the first user-selected characteristics.
\item Print a box plot, line plots (raw & sqrt scale), and bar plots (raw & sqrt scale).
}
}
\examples{
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
exp_data <- DE_exp_data
lipid_char_table <- DE_lipid_char_table
group_info <- DE_group_info
char_var <- colnames(lipid_char_table)[-1]
DE.sub.char.2 <- DE_sub_char_2(exp_data, data_transform=TRUE,
                               lipid_char_table = lipid_char_table,
                               split_var = char_var[1],
                               char_var = char_var[4],
                               group_info = group_info,paired = FALSE,
                               sig_pvalue=0.05, sig_FC=2,
                               exclude_var_missing=TRUE,
                               missing_pct_limit=50, replace_zero=TRUE,
                               zero2what='min', xmin=0.5,replace_NA=TRUE,
                               NA2what='min', ymin=0.5,
                               pct_transform=TRUE, trans_type='log',
                               centering=FALSE, scaling=FALSE)
char.class <- unique(DE.sub.char.2[[2]][1])
DE_sub_char_plot_2(DE.sub.char.2[[2]],
                   DE_split_char_index = DE.sub.char.2[[3]],
                   group_info = group_info, char_var = char_var[4],
                   split_var = char_var[1],
                   split_class = char.class[5,],
                   insert_ref_group=NULL, ref_group=NULL)
}
