% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deChar_multiGroup.R
\name{deChar_multiGroup}
\alias{deChar_multiGroup}
\title{deChar_multiGroup}
\usage{
deChar_multiGroup(
  processed_se,
  char,
  ref_group,
  post_hoc = c("One-way ANOVA", "Kruskal–Wallis test"),
  post_hoc_sig = c("pval", "padj"),
  post_hoc_p_cutoff = 0.05,
  transform = c("none", "log10", "square", "cube", "log2")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the deChar list
returned by \code{\link{list_lipid_char}}.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{post_hoc}{Character. The method to statistic calculation. Allowed method
include "One-way ANOVA" and "Kruskal–Wallis test". Default is \code{'One-way ANOVA'}.}

\item{post_hoc_sig}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{post_hoc_p_cutoff}{Numeric. Significant level. Default is \code{0.05}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube", "log2". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function is for multi-group lipid characteristics analysis.
It identifies the differentially expressed lipid characteristics based on the user-selected characteristic.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_list <- list_lipid_char(processed_se)$deChar_list
print(char_list)
deChar_se <- deChar_multiGroup(
    processed_se, char='class', ref_group='ctrl', post_hoc='One-way ANOVA',
    post_hoc_sig='pval', post_hoc_p_cutoff=0.05, transform='log10')
}
