% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deChar_twoGroup.R
\name{deChar_twoGroup}
\alias{deChar_twoGroup}
\title{deChar_twoGroup}
\usage{
deChar_twoGroup(
  processed_se,
  char,
  ref_group,
  test = c("t-test", "Wilcoxon test"),
  significant = c("pval", "padj"),
  p_cutoff = 0.05,
  FC_cutoff = 1,
  transform = c("none", "log10", "square", "cube", "log2")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the deChar list
returned by \code{\link{list_lipid_char}}.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{test}{Character. The method to use for comparing means.
Allowed method include "t-test" and "Wilcoxon test". Default is \code{'t-test'}.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. Significant level. Default is \code{0.05}.}

\item{FC_cutoff}{Numeric. Significance of the fold-change. Default is \code{1}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube", "log2". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function is for two-group lipid characteristics analysis.
It identifies the differentially expressed lipid characteristics based on the user-selected characteristic.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_list <- list_lipid_char(processed_se)$deChar_list
names(char_list) <- NULL
deChar_se <- deChar_twoGroup(processed_se, char=char_list[14], ref_group="ctrl",
    test='t-test', significant="pval", p_cutoff=0.05,
    FC_cutoff=1, transform='log10')

}
