% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSp_multiGroup.R
\name{deSp_multiGroup}
\alias{deSp_multiGroup}
\title{deSp_multiGroup}
\usage{
deSp_multiGroup(
  processed_se,
  ref_group,
  test = c("One-way ANOVA", "Kruskal-Wallis test"),
  significant = c("pval", "padj"),
  p_cutoff = 0.05,
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{test}{Character. The method to use for comparing means. Allowed method include "One-way ANOVA" and
"Kruskal-Wallis test". Default is \code{'One-way ANOVA'}.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. Significant level. (default: 0.05)}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
Compute differentially expressed analysis of multiple groups
(independent) to find significant lipid species.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deSp_se <- deSp_multiGroup(processed_se, ref_group='ctrl', test='One-way ANOVA',
    significant='pval', p_cutoff=0.05, transform='log10')
}
