% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dim_process.R
\name{dim_process}
\alias{dim_process}
\title{dim_process}
\usage{
dim_process(
  exp_transform_SE,
  sig_feature = NULL,
  type = "PLSDA",
  insert_ref_group = NULL,
  ref_group = NULL
)
}
\arguments{
\item{exp_transform_SE}{A SummarizedExperiment object contains information
about various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
allowed. The row data  corresponds to specific lipid features, such as class
and total length. The first column's name must be "feature" (lipid species),
and NAs are allowed for this data. The column data comprises sample names,
sample labels, group names, and pair numbers that represent 'the pair' for
conducting t-tests or Wilcoxon tests. NAs are allowed.}

\item{sig_feature}{A character string for identifying significant variable in
differential analysis results. It is only required when \bold{type}
is "PLSDA."}

\item{type}{A character of \bold{"PCA"}, \bold{"tsne"}, \bold{"UMAP"},
and \bold{"PLSDA"}. The dimension reduction method to which the transformed
data will be applied.}

\item{insert_ref_group}{A character string. The name of 'ctrl' after name
conversion.}

\item{ref_group}{A character string. The name of 'exp' after name conversion.}
}
\value{
Return a SummarizedExperiment object.
}
\description{
This function transforms the data into a format suitable for
further use in dimension reduction methods, such as PCA, t-SNE, UMAP,
and PLSDA.
}
\examples{
data("DE_data")
exp_data_SE <- DE_data
exp_transform_SE <- data_process(exp_data_SE, exclude_var_missing=TRUE,
    missing_pct_limit=50, replace_zero=TRUE, zero2what='min', xmin=0.5,
    replace_NA=TRUE, NA2what='min', ymin=0.5,  pct_transform=TRUE,
    data_transform=FALSE, trans_type='log', centering=FALSE, scaling=FALSE)
dim_process(exp_transform_SE, sig_feature=NULL,
    type='PCA', insert_ref_group=NULL, ref_group=NULL)
}
