% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{dr_tsne}
\alias{dr_tsne}
\title{dr_tsne}
\usage{
dr_tsne(
  processed_se,
  pca = TRUE,
  perplexity = 5,
  max_iter = 500,
  clustering = c("kmeans", "kmedoids", "hclustering", "dbscan", "group_info"),
  cluster_num = 2,
  kmedoids_metric = NULL,
  distfun = NULL,
  hclustfun = NULL,
  eps = NULL,
  minPts = NULL
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.
(NOTE: A SummarizedExperiment object generated by \code{\link{deSp_twoGroup}},
\code{\link{deChar_twoGroup}}, \code{\link{deSp_multiGroup}}, or
\code{\link{deChar_multiGroup}} is also allowed.)}

\item{pca}{Logical. If pca=TRUE, an initial PCA step will be performed. Default is \code{TRUE}.}

\item{perplexity}{Numeric. Perplexity parameter (should not be bigger than 3 * perplexity < nrow(X) - 1.}

\item{max_iter}{Integer. Number of iterations.}

\item{clustering}{Character. The method to be used for clustering. Allowed method
include "kmeans", "kmedoids", "hclustering", "dbscan", "group_info". Default is \code{"kmeans"}.}

\item{cluster_num}{Numeric. A positive integer specifying the number of clusters.
The number must be between 1 and 10. Default is \code{2}.}

\item{kmedoids_metric}{Character. The metric to be used for calculating
dissimilarities between observations when choosing \code{"kmedoids"} as clustering method.
Must be one of "euclidean" and "manhattan". If "kmedoids" is not selected as
the clustering method, set the value to NULL.}

\item{distfun}{Character. The distance measure to be used when choosing \code{"hclustering"}
as clustering method. Allow method include "pearson", "kendall", "spearman",
"euclidean", "manhattan", "maximum", "canberra", "binary", and "minkowski".
If "hclustering" is not selected as the clustering method, set the value to NULL.}

\item{hclustfun}{Character. The agglomeration method to be used when choosing
\code{"hclustering"} as clustering method. This should be (an unambiguous abbreviation of)
one of "ward.D", "ward.D2", "single", "complete", "average" (=UPGMA),
"mcquitty" (= WPGMA), "median" (= WPGMC), or "centroid" (= UPGMC).
If "hclustering" is not selected as the clustering method, set the value to NULL.}

\item{eps}{Numeric. The size of the epsilon neighborhood when choosing \code{"dbscan"} as clustering method.
If "dbscan" is not selected as the clustering method, set the value to NULL.}

\item{minPts}{Numeric. The number of minimum points in the eps region (for core points)
when choosing \code{"dbscan"} as clustering method.
If "dbscan" is not selected as the clustering method, set the value to NULL.}
}
\value{
Return a list with 1 data frame, 1 interactive plot, and 1 static plot.
\enumerate{
\item tsne_result: a data frame of tsne data.
\item interactive_tsne & static_tsne: tsne plot.
}
}
\description{
t-distributed stochastic neighbor embedding (t-SNE) is a dimensionality reduction
method that transforms data from a high-dimensional space into a low-dimensional
space while retaining the original data's essential properties.
}
\examples{
data("profiling_data")
processed_se <- data_process(profiling_data, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5, normalization='Percentage')
result_tsne <- dr_tsne(processed_se, pca=TRUE, perplexity=5, max_iter=500, clustering='kmeans',
    cluster_num=2, kmedoids_metric=NULL, distfun=NULL, hclustfun=NULL, eps=NULL, minPts=NULL)
}
