% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimension_reduction.R
\name{dr_umap}
\alias{dr_umap}
\title{dr_umap}
\usage{
dr_umap(
  processed_se,
  n_neighbors = 15,
  scaling = c("none", FALSE, NULL, "Z", "scale", TRUE, "maxabs", "range", "colrange"),
  umap_metric = c("euclidean", "cosine", "manhattan", "hamming", "categorical"),
  clustering = c("kmeans", "kmedoids", "hclustering", "dbscan", "group_info"),
  cluster_num = 2,
  kmedoids_metric = NULL,
  distfun = NULL,
  hclustfun = NULL,
  eps = NULL,
  minPts = NULL
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{n_neighbors}{Numeric. The size of local neighborhood (in terms of number of
neighboring sample points) used for manifold approximation.}

\item{scaling}{Logical/Character. Scaling to apply to \bold{X} if it is a data frame or matrix:
\itemize{
\item \code{"none"} or \code{FALSE} or \code{NULL}: No scaling.
\item \code{"Z"} or \code{"scale"} or \code{TRUE}: Scale each column to
zero mean and variance 1.
\item \code{"maxabs"}: Center each column to mean 0, then divide each
element by the maximum absolute value over the entire matrix.
\item \code{"range"}: Range scale the entire matrix, so the smallest
element is 0 and the largest is 1.
\item \code{"colrange"}: Scale each column in the range (0,1).
}
Default is \code{TRUE}.}

\item{umap_metric}{Character. Type of distance metric to use to find nearest neighbors.
One of "euclidean", "cosine", "manhattan", "hamming", "categorical". Default is \code{'euclidean'}.}

\item{clustering}{Character. The method to be used for clustering. Allowed method
include "kmeans", "kmedoids", "hclustering", "dbscan", "group_info". Default is \code{"kmeans"}.}

\item{cluster_num}{Numeric. A positive integer specifying the number of clusters.
The number must be between 1 and 10. Default is \code{2}.}

\item{kmedoids_metric}{Character. The metric to be used for calculating
dissimilarities between observations when choosing \code{"kmedoids"} as clustering method.
Must be one of "euclidean" and "manhattan". If "kmedoids" is not selected as
the clustering method, set the value to NULL.}

\item{distfun}{Character. The distance measure to be used when choosing \code{"hclustering"}
as clustering method. Allow method include "pearson", "kendall", "spearman",
"euclidean", "manhattan", "maximum", "canberra", "binary", and "minkowski".
If "hclustering" is not selected as the clustering method, set the value to NULL.}

\item{hclustfun}{Character. The agglomeration method to be used when choosing
\code{"hclustering"} as clustering method. This should be (an unambiguous abbreviation of)
one of "ward.D", "ward.D2", "single", "complete", "average" (=UPGMA),
"mcquitty" (= WPGMA), "median" (= WPGMC), or "centroid" (= UPGMC).
If "hclustering" is not selected as the clustering method, set the value to NULL.}

\item{eps}{Numeric. The size of the epsilon neighborhood when choosing \code{"dbscan"} as clustering method.
If "dbscan" is not selected as the clustering method, set the value to NULL.}

\item{minPts}{Numeric. The number of minimum points in the eps region (for core points)
when choosing \code{"dbscan"} as clustering method.
If "dbscan" is not selected as the clustering method, set the value to NULL.}
}
\value{
Return a list with 1 data frame, 1 interactive plot, and 1 static plot.
\enumerate{
\item umap_result: a data frame of UMAP data.
\item interactive_umap & static_umap: UMAP plot
}
}
\description{
Uniform Manifold Approximation and Projection (UMAP) is a dimensionality reduction
method that transforms data from a high-dimensional space into a low-dimensional
space while retaining the original data's essential properties.
}
\examples{
data("profiling_data")
processed_se <- data_process(profiling_data, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5, normalization='Percentage')
result_umap <- dr_umap(processed_se, n_neighbors=15, scaling=TRUE, umap_metric='euclidean',
    clustering='kmeans', cluster_num=2, kmedoids_metric=NULL,
    distfun=NULL, hclustfun=NULL, eps=NULL, minPts=NULL)
}
