% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_ora.R
\name{enrichment_ora}
\alias{enrichment_ora}
\title{enrichment_ora}
\usage{
enrichment_ora(
  deSp_se,
  char = NULL,
  significant = c("pval", "padj"),
  p_cutoff = 0.05
)
}
\arguments{
\item{deSp_se}{The resulting SummarizedExperiment object from the differential
expression analysis function, such as \code{\link{deSp_twoGroup}} and
\code{\link{deSp_multiGroup}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}. Enter NULL to plot all characteristics.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. The threshold to distinguish enriched lipid-sets
from not-enriched ones. Default is \code{0.05}.}
}
\value{
Return a list of enrichment result, 1 interactive plot, 1 static plot, and 1 table.
\enumerate{
\item enrich_result: a table of enrichment result.
\item static_barPlot: a static bar plot that classifies significant lipid species into 'up-regulated' or 'down-regulated' categories.
\item interactive_barPlot: an interactive bar plot that classifies significant lipid species into 'up-regulated' or 'down-regulated' categories.
\item table_barPlot: table for plotting bar plots.
}
}
\description{
This function conducts Over-Representation Analysis (ORA) to
determine whether significant lipid species are enriched in specific lipid class categories.
}
\examples{
data("de_data_twoGroup")
processed_se_twoGroup <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
deSp_se_twoGroup <- deSp_twoGroup(
    processed_se_twoGroup, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
ora_all_twoGroup <- enrichment_ora(
    deSp_se_twoGroup, char=NULL, significant='pval', p_cutoff=0.05)
char_list <- list_lipid_char(processed_se_twoGroup)$common_list
print(char_list)
ora_one_twoGroup <- enrichment_ora(
    deSp_se_twoGroup, char='class', significant='pval', p_cutoff=0.05)

data("se_multiGroup")
processed_se_multiGroup <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
deSp_se_multiGroup <- deSp_multiGroup(
    processed_se_multiGroup, ref_group='ctrl', test='One-way ANOVA',
    significant='pval', p_cutoff=0.05, transform='log10')
ora_all_multiGroup <- enrichment_ora(
    deSp_se_multiGroup, char=NULL, significant='pval', p_cutoff=0.05)
char_list <- list_lipid_char(processed_se_multiGroup)$common_list
print(char_list)
ora_one_multiGroup <- enrichment_ora(
    deSp_se_multiGroup, char='class', significant='pval', p_cutoff=0.05)
}
