% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalution_plot.R
\name{evalution_plot}
\alias{evalution_plot}
\title{evalution_plot}
\usage{
evalution_plot(data, method)
}
\arguments{
\item{data}{confusion matrix after cross-validation. A output data frame of \code{\link{ML_final}} (list[[2]],cv_cm_result).}

\item{method}{A character string indicating which evaluation method to be used. Allowed methods include "Accuracy", "Sensitivity", "Specificity", "Positive Predictive Value", "Negative Predictive Value", "Precision", "Recall", "F1", "Prevalence", "Detection Rate", "Detection Prevalence", "Balance Accuracy".}
}
\value{
Return a list of 1 tibble and 1plot.
\enumerate{
\item evalution_data: a tibble of model evaluation information.
\item evalution_plot: model performance plot
}
}
\description{
This function provides useful indicators for users to evaluate model performance, such as the average value and 95% confidence interval of accuracy, sensitivity (recall), specificity, positive predictive value (precision), negative predictive value, F1 score, prevalence, detection rate, detection prevalence, balanced accuracy.
}
\examples{
\dontrun{
data("ML_exp_data")
data("ML_lipid_char_table")
data("ML_condition_table")
exp_data <- ML_exp_data
lipid_char_table <- ML_lipid_char_table
condition_table <- ML_condition_table
char_var <- colnames(lipid_char_table)[-1]
ML_data <- ML_data_process(exp_data, group_info = condition_table, lipid_char_table,
                           char_var[1], exclude_var_missing=T, missing_pct_limit=50,
                           replace_zero=T, zero2what='min', xmin=0.5,replace_NA=T,
                           NA2what='min', ymin=0.5, pct_transform=T, data_transform=T,
                           trans_type='log', centering=F, scaling=F)
ML_output <- ML_final(ML_data[[2]], ranking_method='Random_forest', ML_method='Random_forest',
                      split_prop=0.3, nfold=10)
evalution_plot(ML_output[[2]], method='Accuracy')
}
}
