% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_profiling.R
\name{exp_profiling}
\alias{exp_profiling}
\title{exp_profiling}
\usage{
exp_profiling(exp_data)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}
}
\value{
Return 3 plots.
\enumerate{
\item histogram of number of expressed lipids.
\item histogram of the total amount of lipid in each sample.
\item density plot of the underlying probability distribution of the lipid expression in each sample.
}
}
\description{
This function three types of distribution plots for users to a simple view of sample variability, and compare the amount/expression difference of lipid between samples (i.e., patients vs. control).
\enumerate{
\item The first histogram depicts the numbers of lipids expressed in each sample.
\item The second histogram illustrates the total amount of lipid in each sample.
\item The last density plot visualizes the underlying probability distribution of the lipid expression in each sample (line).
}
}
\examples{
data("profiling_exp_data")
exp_data <- profiling_exp_data
exp_profiling(exp_data)
}
