% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-df.R
\name{extract_summarized_experiment}
\alias{extract_summarized_experiment}
\title{extract_summarized_experiment}
\usage{
extract_summarized_experiment(se)
}
\arguments{
\item{se}{A SummarizedExperiment object.}
}
\value{
Return a list of data frames.
}
\description{
This function returns a list of data frames includes abundance data, lipid
characteristic table, group information table, and other analysis results from
a SummarizedExperiment object.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
deSp_se <- deSp_twoGroup(processed_se, ref_group='ctrl', test='t-test',
    significant='padj', p_cutoff=0.05, FC_cutoff=2, transform='log10')
table_list <- extract_summarized_experiment(deSp_se)
}
