% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_chain_db.R
\name{heatmap_chain_db}
\alias{heatmap_chain_db}
\title{heatmap_chain_db}
\usage{
heatmap_chain_db(
  processed_se,
  char,
  char_feature = NULL,
  ref_group,
  test = c("t-test", "Wilcoxon test", "One-way ANOVA", "Kruskal-Wallis test"),
  significant = c("pval", "padj"),
  p_cutoff = 0.05,
  FC_cutoff = 1,
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the chain_db list
returned by \code{\link{list_lipid_char}}.}

\item{char_feature}{Character/NULL. A feature selected by users from \code{char}
to visualize the specific plot for the selected category of that characteristic.
For example, if char is 'class' and char_feature is 'Cer', the resulting
plots will display data for 'Cer' within the 'class' category.
Set NULL to prevent selecting any feature as char_feature.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{test}{Character. The method to use for comparing means.
Allowed method include "t-test", "Wilcoxon test", "One-way ANOVA", and "Kruskal-Wallis test".
"t-test", "Wilcoxon test" are for two-group data, and "One-way ANOVA"
and "Kruskal-Wallis test" are for multi-group data. Default is \code{'t-test'}.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. Significant level. Default is \code{0.05}.}

\item{FC_cutoff}{Numeric. Significance of the fold-change, which is only
applicable for the two-group data. Default is \code{1}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a list of 2 lists.
\enumerate{
\item total_chain: the result list of total chain.
\item each_chain: the result list of fatty acids chain.
}
}
\description{
This function generates a heatmap showing the correlation
between the double bond and chain length of lipids.
}
\examples{
data("de_data_twoGroup")
processed_se_twoGroup <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
char_list <- list_lipid_char(processed_se_twoGroup)$chain_db_list
print(char_list)
heatmap_all_twoGroup <- heatmap_chain_db(
    processed_se_twoGroup, char='class', char_feature=NULL, ref_group='ctrl',
    test='t-test', significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
heatmap_one_twoGroup <- heatmap_chain_db(
    processed_se_twoGroup, char='class', char_feature='PC', ref_group='ctrl',
    test='t-test', significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')

data("se_multiGroup")
processed_se_multiGroup <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
char_list <- list_lipid_char(processed_se_multiGroup)$chain_db_list
print(char_list)
heatmap_all_multiGroup <- heatmap_chain_db(
    processed_se_multiGroup, char='class', char_feature=NULL, ref_group='ctrl',
    test='One-way ANOVA', significant='pval', p_cutoff=0.05, FC_cutoff=NULL,
    transform='log10')
heatmap_one_multiGroup <- heatmap_chain_db(
    processed_se_multiGroup, char='class', char_feature='PC', ref_group='ctrl',
    test='One-way ANOVA', significant='pval', p_cutoff=0.05, FC_cutoff=NULL,
    transform='log10')
}
