% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_clustering.R
\name{heatmap_clustering}
\alias{heatmap_clustering}
\title{heatmap_clustering}
\usage{
heatmap_clustering(
  de_se,
  char,
  distfun = c("pearson", "kendall", "spearman"),
  hclustfun = c("ward.D", "ward.D2", "single", "complete", "average", "mcquitty",
    "median", "centroid"),
  type = c("all", "sig")
)
}
\arguments{
\item{de_se}{The resulting SummarizedExperiment object from the differential
expression analysis function, such as \code{\link{deSp_twoGroup}}, \code{\link{deSp_multiGroup}},
\code{\link{deChar_twoGroup}}, and \code{\link{deChar_multiGroup}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}.}

\item{distfun}{Character. The distance measure for computing correlation coefficient
(or covariance). Allowed methods include "pearson", "kendall", and "spearman".
Default is \code{'pearson'}}

\item{hclustfun}{Character. The agglomeration method. This should be
(an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete",
"average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC), or "centroid" (= UPGMC).
Default is \code{'complete'}}

\item{type}{Character. It must be 'all' or 'sig.' 'all' for output the results of all
lipid species or characteristics, and 'sig' for output the results of
significant lipid species or characteristics.}
}
\value{
Return a list with 1 figure and 1 matrix.
\enumerate{
\item interactive_heatmap & static_heatmap: a heatmap provides an overview of
all/significant lipid species or characteristics that illustrates the differences between groups.
\item corr_coef_matrix: the matrix of the heatmap.
}
}
\description{
Hierarchical clustering of lipid species or characteristics
derived from two/multiple groups.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deSp_se <- deSp_twoGroup(processed_se, ref_group='ctrl', test='t-test',
    significant='padj', p_cutoff=0.05, FC_cutoff=2, transform='log10')
result <- heatmap_clustering(de_se=deSp_se, char='class', distfun='pearson',
    hclustfun='complete', type='sig')
}
