% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_annotation.R
\name{lipid_annotation}
\alias{lipid_annotation}
\title{lipid_annotation}
\usage{
lipid_annotation(goslin_annotation)
}
\arguments{
\item{goslin_annotation}{A data frame of lipid characteristics from the
\code{\link[rgoslin]{parseLipidNames}} function of \pkg{\link{rgoslin}} package.}
}
\value{
Return a data frame of lipid characteristics table.
}
\description{
This function combines adjusting the annotation table
returned by Goslin with adding mappings between lipids and the LION ontology
and mappings with other resource IDs. The returned lipid characteristics table
will be used for further analysis in LipidSigR.
}
\examples{
library(dplyr)
data("abundance_twoGroup")
parse_lipid <- rgoslin::parseLipidNames(abundance_twoGroup$feature)
recognized_lipid <- parse_lipid$Original.Name[which(parse_lipid$Grammar != 'NOT_PARSEABLE')]
goslin_annotation <- parse_lipid \%>\%
    dplyr::filter(Original.Name \%in\% recognized_lipid)
lipid_annotation_table <- lipid_annotation(goslin_annotation)
}
