% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_profiling.R
\name{lipid_profiling}
\alias{lipid_profiling}
\title{lipid_profiling}
\usage{
lipid_profiling(processed_se, char)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the common list r
eturned by \code{\link{list_lipid_char}}.}
}
\value{
Return a list of 2 interactive plots, 2 static plots, and 2 tables.
\enumerate{
\item interactive_char_barPlot & static_char_barPlot: bar plot classified by
selected lipid characteristic.
within each group (e.g., PE, PC) of selected characteristics (e.g., class).
\item interactive_lipid_composition & static_lipid_composition: stacked horizontal
bar chart of lipid class composition.
\item table_char_barPlot: table for plotting char_barPlot.
\item table_lipid_composition: table for plotting lipid_composition.
}
}
\description{
This function provides two plots for users to explore lipid
abundance over a specific lipid characteristic.
}
\examples{
data("profiling_data")
processed_se <- data_process(
    profiling_data, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
list_lipid_char(processed_se)$common_list
result <- lipid_profiling(processed_se, char="class")
}
