\name{lipidChar}
\alias{lipidChar}
\docType{data}
\title{
lipidChar
}
\description{
A data frame of lipid characteristics.
The table with each characteristic's aspect, characteristic, and source.
}
\usage{data("lipidChar")}
\format{
  A data frame with 25 observations on the following 3 variables.
  \describe{
    \item{\code{aspect}}{a character vector}
    \item{\code{characteristic}}{a character vector}
    \item{\code{source}}{a character vector}
  }
}
\source{
J Biochem. 2012 Nov;152(5):387-95., J Lipid Res. 2016 Jun;57(6):943-54.
}
\examples{
data(lipidChar)
}
\keyword{datasets}
