% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-df.R
\name{list_lipid_char}
\alias{list_lipid_char}
\title{list_lipid_char}
\usage{
list_lipid_char(processed_se)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object after conducting \code{\link{data_process}}.}
}
\value{
Return a list of named vector.
\enumerate{
\item deChar_list: List of lipid characteristics for conducting differential
expressed lipid characteristics analysis.
\item chain_db_list: List of lipid characteristics for performing chain
length-double bond correlation analysis.
\item common_list: List of lipid characteristics for other analyses in
LipidSigR, such as Profiling and Enrichment.
}
}
\description{
This function lists the lipid characteristics suitable for
different types of analyses.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_list <- list_lipid_char(processed_se)
}
