\name{ML_condition_table}
\alias{ML_condition_table}
\docType{data}
\title{
ML_condition_table
}
\description{
A data frame of sample names and clinical conditions for machine learning analysis workflow.
}
\usage{data("ML_condition_table")}
\format{
  A data frame with 228 observations on the following 2 variables.
  \describe{
    \item{\code{sample_name}}{a character vector}
    \item{\code{group}}{a numeric vector}
  }
}
\examples{
data(ML_condition_table)
## maybe str(ML_condition_table) ; plot(ML_condition_table) ...
}
\keyword{datasets}
