% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_corr_network.R
\name{ml_corr_network}
\alias{ml_corr_network}
\title{ml_corr_network}
\usage{
ml_corr_network(
  ml_se,
  feature_importance = "Algorithm-based",
  correlation = "pearson",
  edge_cutoff = 0,
  feature_num = 10,
  nsim = 5
)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_importance}{Character. The method of feature importance. Allowed methods are
'Algorithm-based' and 'SHAP'. Default is \code{'Algorithm-based'}.}

\item{correlation}{Character. The method for computing correlation coefficient.
Allowed methods includes 'pearson', 'kendall', and 'spearman'. Default is \code{'pearson'}.}

\item{edge_cutoff}{Numeric. A value between 0 and 1. Only the correlation
coefficient larger than it will be shown as a line in the plot.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}

\item{nsim}{Integer. The times of simulation. Default is \code{5}.}
}
\value{
Return 1 interactive plot, 1 static plot, and 1 table.
\enumerate{
\item interactive_correlation_network & static_correlation_network: correlation network.
\item edge_table & node_table: table for plotting network.
}
}
\description{
Select feature importance from the result of \code{\link{ml_model}}.
Two methods are available to rank feature importance: algorithm-based and SHAP
analysis. The average importance of the top features across all CV runs will
be displayed by specifying a feature count.
}
\examples{
data("ml_se_sub")
res <- ml_corr_network(ml_se_sub, feature_importance='Algorithm-based',
    correlation='pearson', edge_cutoff=0, feature_num=10, nsim=5)

}
