% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ML_data_process.R
\name{ML_data_process}
\alias{ML_data_process}
\title{ML_data_process}
\usage{
ML_data_process(
  exp_data,
  group_info,
  lipid_char_table,
  char_var = NULL,
  exclude_var_missing = T,
  missing_pct_limit = 50,
  replace_zero = T,
  zero2what = "min",
  xmin = 0.5,
  replace_NA = T,
  NA2what = "min",
  ymin = 0.5,
  pct_transform = T,
  data_transform = T,
  trans_type = "log",
  centering = F,
  scaling = F
)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}

\item{group_info}{A data frame of one clinical term (or one set of clinical terms) from demographic data.  The first column is the sample name and the second column is the group number assigned to each sample. The group number can only be 0 or 1. NAs are not allowed.}

\item{lipid_char_table}{A data frame of lipid characteristics such as name(feature) of lipid, class of lipid, the total length of lipid, and Fatty acid (FA_) related characteristics. NAs are allowed. The name of the first column must be "feature" (lipid species).}

\item{char_var}{A character string of the lipid characteristic selected by users from the column name of \bold{lipid_char_table}, such as 'class', 'total length'.}

\item{exclude_var_missing}{Logical. Remove Lipid with too many missing values. (default: T)}

\item{missing_pct_limit}{An integer indicating the missing values over a certain percentage should be removed.}

\item{replace_zero}{Logical. Replace 0. (default: T)}

\item{zero2what}{A character string indicating the value to replace 0. Value include 'mean', 'median', 'min'. (default: min)}

\item{xmin}{A numeric value indicating the min value to replace 0.}

\item{replace_NA}{Logical. If remove_na = TRUE, all NA will be removed. (default: T)}

\item{NA2what}{A character string indicating the value to replace NA. Value include 'mean', 'median', 'min'. (default: min)}

\item{ymin}{A numeric value indicating the min value to replace NA.}

\item{pct_transform}{Logical. Transform lipid value into a percentage. (default:T)}

\item{data_transform}{Logical. Data transformation(log). (default:T)}

\item{trans_type}{A character string of data transformation type. (default:'log')}

\item{centering}{A logical value indicating whether the variables should be shifted to be zero centered. Alternately, a vector of length equal the number of columns of x can be supplied. The value is passed to scale. (default: F)}

\item{scaling}{A logical value. If scaling = TRUE, each block is standardized to zero means and unit variances. (default: F).}
}
\value{
Return a list of 2 data frames.
\enumerate{
\item [[1]]: the processed data can be used as input of machine learning analysis. The value of this data frame is adding the lipid (feature)'s original expression to the expression of user-selected lipid characteristics.
\item [[2]]: The transpose data frame of \bold{[[1]]}, which is also processed data and can be used as the input of machine learning analysis. The value of this data frame is adding the lipid (feature)'s original expression to the expression of user-selected lipid characteristics.
}
}
\description{
This function provides data preprocessing and output processed data for machine learning.
}
\examples{
\dontrun{
data("ML_exp_data")
data("ML_lipid_char_table")
data("ML_condition_table")
exp_data <- ML_exp_data
lipid_char_table <- ML_lipid_char_table
condition_table <- ML_condition_table
char_var <- colnames(lipid_char_table)[-1]
ML_data_process(exp_data, group_info = condition_table, lipid_char_table, char_var[1],
                exclude_var_missing=T, missing_pct_limit=50, replace_zero=T,
                zero2what='min', xmin=0.5,replace_NA=T, NA2what='min', ymin=0.5,
                pct_transform=T, data_transform=T, trans_type='log',
                centering=F, scaling=F)
}
}
