\name{ML_lipid_char_table}
\alias{ML_lipid_char_table}
\docType{data}
\title{
ML_lipid_char_table
}
\description{
A data frame of lipid characteristics for machine learning analysis workflow.
}
\usage{data("ML_lipid_char_table")}
\format{
  A data frame with 89 observations on the following 6 variables.
  \describe{
    \item{\code{feature}}{a character vector}
    \item{\code{class}}{a character vector}
    \item{\code{structural_category}}{a character vector}
    \item{\code{functional_category}}{a character vector}
    \item{\code{totallength}}{a numeric vector}
    \item{\code{totaldb}}{a numeric vector}
  }
}
\examples{
data(ML_lipid_char_table)
}
\keyword{datasets}
