% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{ml_shap}
\alias{ml_shap}
\title{ml_shap}
\usage{
ml_shap(ml_se, feature_num = 10, nsim = 5)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}

\item{nsim}{Numeric. A positive integer indicating the times of simulation.
Default is \code{5}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function utilizes the Shapley Additive exPlanations (SHAP)
method to rank feature importance based on a user-defined number of features.
}
\examples{
data("ml_se_sub")
shap_se <- ml_shap(ml_se_sub, feature_num=10, nsim=5)
}
