% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_gatom.R
\name{nw_gatom}
\alias{nw_gatom}
\title{nw_gatom}
\usage{
nw_gatom(
  deSp_se,
  organism = c("human", "mouse"),
  n_lipid = 50,
  sp_significant = c("pval", "padj"),
  sp_p_cutoff = 0.05,
  sp_FC_cutoff = 1
)
}
\arguments{
\item{deSp_se}{A SummarizedExperiment object with results computed by \code{\link{deSp_twoGroup}}.}

\item{organism}{Character. The species to which the genes will be matched.
Allowed species are "human" and "mouse". Default is \code{'human'}.}

\item{n_lipid}{Numeric. The number of lipids to be scored positively determines
the size of the resulting module. The higher the number, the larger the module.
Default is \code{50}.}

\item{sp_significant}{Character. The p-value to be used for the statistically
significant of lipid species. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{sp_p_cutoff}{Numeric. Significant level of lipid species. Default is \code{0.05}.}

\item{sp_FC_cutoff}{Numeric. Significance of the fold-change of lipid species. Default is \code{1}.}
}
\value{
Return a list of 4 tables.
\enumerate{
\item table_edge: a table of network edges.
\item table_node; a table of network nodes.
\item table_reaction: a table of reactions.
\item table_stat: a table of statistical results.
}
}
\description{
This function returns node and edge tables that can be used to
generate a network using the network visualization tool. It isolates
significant subnetworks within a constructed metabolite-level network.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deSp_se <- deSp_twoGroup(
    processed_se, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
network_table <- nw_gatom(
    deSp_se, organism='mouse', n_lipid=50, sp_significant='pval',
    sp_p_cutoff=0.05, sp_FC_cutoff=1)
}
