% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_lipid_reaction.R
\name{nw_lipid_reaction}
\alias{nw_lipid_reaction}
\title{nw_lipid_reaction}
\usage{
nw_lipid_reaction(
  deSp_se,
  organism = c("human", "mouse"),
  show_sp = c("all", "sigClass", "none"),
  show_all_reactions = FALSE,
  sp_significant = c("pval", "padj"),
  sp_p_cutoff = 0.05,
  sp_FC_cutoff = 1,
  class_significant = c("pval", "padj"),
  class_p_cutoff = 0.05,
  class_FC_cutoff = 1
)
}
\arguments{
\item{deSp_se}{A SummarizedExperiment object with results computed by \code{\link{deSp_twoGroup}}.}

\item{organism}{Character. The species to which the genes will be matched.
Allowed species are "human" and "mouse". Default is \code{'human'}.}

\item{show_sp}{Character. Determine how lipid species around the lipid class will be displayed.
Must be one of "all", "sigClass", and "none". Select "all" to show all species,
"sigClass" to show species in significant lipid classes, and "none" to not show any species.
Default is \code{'sigClass'}.}

\item{show_all_reactions}{Logical. If show_all_reactions=TURE, all the reactions will be showed.
Default is \code{FALSE}.}

\item{sp_significant}{Character. The p-value to be used for the statistically
significant of lipid species. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{sp_p_cutoff}{Numeric. Significant level of lipid species. Default is \code{0.05}.}

\item{sp_FC_cutoff}{Numeric. Significance of the fold-change of lipid species. Default is \code{1}.}

\item{class_significant}{Character. The p-value to be used for the statistically
significant of lipid class. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{class_p_cutoff}{Numeric. Significant level of lipid class. Default is \code{0.05}.}

\item{class_FC_cutoff}{Numeric. Significance of the fold-change of lipid class. Default is \code{1}.}
}
\value{
Return a list of 4 tables.
\enumerate{
\item table_edge: a table of network edges.
\item table_node; a table of network nodes.
\item table_reaction: a table of reactions.
\item table_stat: a table of statistical results.
}
}
\description{
This function returns node and edge tables that can be used to
generate a network using the network visualization tool. It is designed to
graphically represent significant lipid classes and species within lipid
biosynthesis pathways.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deSp_se <- deSp_twoGroup(
    processed_se, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
network_table <- nw_lipid_reaction(
    deSp_se, organism='mouse', show_sp='sigClass', show_all_reactions=FALSE,
    sp_significant='pval', sp_p_cutoff=0.05, sp_FC_cutoff=1,
    class_significant='pval', class_p_cutoff=0.05, class_FC_cutoff=1)
}
