% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nw_pathway_activity.R
\name{nw_pathway_activity}
\alias{nw_pathway_activity}
\title{nw_pathway_activity}
\usage{
nw_pathway_activity(deSp_se, organism = c("human", "mouse"))
}
\arguments{
\item{deSp_se}{A SummarizedExperiment object with results computed by \code{\link{deSp_twoGroup}}.}

\item{organism}{Character. The species to which the genes will be matched.
Allowed species are "human" and "mouse". Default is \code{'human'}.}
}
\value{
Return a list of 4 tables.
\enumerate{
\item table_edge: a table of network edges.
\item table_node; a table of network nodes.
\item table_pathway_score: a table of pathway score
\item table_zScore: a table of z score.
}
}
\description{
This function returns node and edge tables that can be used to
generate a network using the network visualization tool. It computes flux
changes in the lipid reaction network, facilitating the identification of
active or suppressed pathways.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
deSp_se <- deSp_twoGroup(
    processed_se, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
network_table <- nw_pathway_activity(deSp_se, organism='mouse')
}
