% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathview_function.R
\name{pathview_function}
\alias{pathview_function}
\title{DE_pathview_function}
\usage{
pathview_function(lipid_class, path, lipid_gene_path, pathway_gene_list)
}
\arguments{
\item{lipid_class}{a vector comprising lipid class, such as PC, PE, TAG}

\item{path}{a character string, output path.}

\item{lipid_gene_path}{A data frame of pathway information, which can be retrieved and reorganized from KEGG or other databases. It includes the related pathways of lipids, ID, pathway ID, pathway name, gene ID, and gene name.}

\item{pathway_gene_list}{A list that comprises the genes included in each pathway for compiling the significant network of the selected lipid class. \emph{NOTE: the information can be retrieved from the KEGG database or others.}}
}
\value{
Return a network plot of pathway analysis and a data frame revealing the details of the plot
\enumerate{
\item pathview plot save as .png
\item A data frame with key_name, DB, ID, path_id, path_name
}
}
\description{
The function is to produce the enrichment plot that assists users to determine whether significant lipid species are enriched in the categories of the selected characteristics. To obtain the enrichment plot, a vector of lipid classes and the pathway information from the database, such as KEGG, need to be provided. A differentially expressed analysis pathview plot and a data frame with key_name, DB, ID, path_id, path_name will be returned as a result.
}
\examples{
\dontrun{
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
data("DE.lipid.gene.path")
data("DE.pathway.gene.list")
exp_data <- DE_exp_data
lipid_char_table <- DE_lipid_char_table
group_info <- DE_group_info
DE.lipid.gene.path <- DE.lipid.gene.path
DE.pathway.gene.list <- DE.pathway.gene.list
exp_transform <- data_process(exp_data, exclude_var_missing=T, missing_pct_limit=50,
                              replace_zero=T, zero2what='min', xmin=0.5, replace_NA=T,
                              NA2what='min', ymin=0.5, pct_transform=T, data_transform=T,
                              trans_type='log', centering=F,  scaling=F)
exp_transform_non_log <- data_process(exp_data, exclude_var_missing=T, missing_pct_limit=50,
                                      replace_zero=T, zero2what='min', xmin=0.5,
                                      replace_NA=T, NA2what='min', ymin=0.5,
                                      pct_transform=T, data_transform=F, trans_type='log',
                                      centering=F,  scaling=F)
lipid_char_filter <- lipid_char_table \%>\% filter(feature \%in\% exp_transform$feature)
DE_species_table_sig <- DE_species_2(exp_transform_non_log, data_transform = T,
                                     group_info = group_info, paired = F,
                                     test = 't.test', adjust_p_method = 'BH',
                                     sig_stat = 'p.adj', sig_pvalue = 0.05,
                                     sig_FC = 2)$DE_species_table_sig
char_var <- colnames(lipid_char_filter)[-1]
enrich_char_table <- Enrichment (DE_species_table_sig, lipid_char_table = lipid_char_filter,
                                 char_var = char_var[1], sig_pvalue = 0.05)$enrich_char_table
sig_enrich_class <- enrich_char_table \%>\% filter(significant == 'YES') \%>\%
                    distinct(characteristic) \%>\% .$characteristic
dir.create(file.path(getwd(),"pathview_result"), recursive=T)
outPath <- file.path(getwd(), "pathview_result")
pathview_function(sig_enrich_class, path = getwd(), DE.lipid.gene.path, DE.pathway.gene.list)
}
}
