% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_process.R
\name{plot_data_process}
\alias{plot_data_process}
\title{plot_data_process}
\usage{
plot_data_process(se, processed_se)
}
\arguments{
\item{se}{A SummarizedExperiment object construct by \code{\link{as_summarized_experiment}}.}

\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}
}
\value{
Return a list with 4 static plots and 4 interactive plots.
\enumerate{
\item static_boxPlot_before: a static bar plot of lipid abundance before data processing.
\item static_densityPlot_before: a static density plot of lipid abundance before data processing.
\item static_boxPlot_after: a static bar plot of lipid abundance before data processing.
\item static_densityPlot_after: a static density plot of lipid abundance after data processing.
\item interactive_boxPlot_before: an interactive bar plot of lipid abundance before data processing.
\item interactive_densityPlot_before: an interactive density plot of lipid abundance before data processing.
\item interactive_boxPlot_after: an interactive bar plot of lipid abundance after data processing.
\item interactive_densityPlot_after: an interactive density plot of lipid abundance after data processing.
}
}
\description{
This function plots the differences in abundance data before and after data processing.
}
\examples{
data("profiling_data")
processed_se <- data_process(
    profiling_data, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
plots <- plot_data_process(profiling_data, processed_se)
}
