% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deChar_multiGroup.R
\name{plot_deChar_multiGroup}
\alias{plot_deChar_multiGroup}
\title{plot_deChar_multiGroup}
\usage{
plot_deChar_multiGroup(deChar_se)
}
\arguments{
\item{deChar_se}{A SummarizedExperiment object with results computed by \code{\link{deChar_multiGroup}}.}
}
\value{
Return a list with 5 static plots, 4 interactive plots, and 5 data frames.
\enumerate{
\item static_barPlot: a static bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic.
\item static_barPlot_sqrt: a static bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_linePlot: a static line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic.
\item static_linePlot_sqrt: a static line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_boxPlot: a static box plot of ctrl group and experiment group.
\item interactive_barPlot: an interactive bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic.
\item interactive_barPlot_sqrt: an interactive bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item interactive_linePlot: an interactive line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic.
\item interactive_linePlot_sqrt: an interactive line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item table_barPlot: table for plotting bar plots
\item table_linePlot: table for plotting line plots
\item table_boxPlot: table for plotting box plots
\item table_char_index: table with the value calculated by the weighted average of lipid characteristics abundance
\item table_index_stat: table with statistics of control and experiment groups
}
}
\description{
This function is for plotting the results of lipid characteristics differential expression analysis.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deChar_se <- deChar_multiGroup(
    processed_se, char='Total.C', ref_group='ctrl', post_hoc='One-way ANOVA',
    post_hoc_sig='pval', post_hoc_p_cutoff=0.05, transform='log10')
deChar_plot <- plot_deChar_multiGroup(deChar_se)
}
