% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_deSp_multiGroup.R
\name{plot_deSp_multiGroup}
\alias{plot_deSp_multiGroup}
\title{plot_deSp_multiGroup}
\usage{
plot_deSp_multiGroup(deSp_se)
}
\arguments{
\item{deSp_se}{A SummarizedExperiment object with results computed by \code{\link{deSp_multiGroup}}.}
}
\value{
Return a list of 1 interactive plot, 1 static plot, and 1 table.
\enumerate{
\item interactive_de_lipid & static_de_lipid: a lollipop chart reveals the top 20 lipid species that pass chosen cut-off.
\item interactive_dotPlot & static_dotPlot: a dot plot reveals the all lipid species that pass chosen cut-off.
\item table_de_lipid: table for plotting DE lipid plot.
\item table_dotPlot: table for plotting dot plot.
}
}
\description{
This function is for plotting the results of lipid species differential expression analysis.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
deSp_se <- deSp_multiGroup(processed_se, ref_group='ctrl', test='One-way ANOVA',
    significant='pval', p_cutoff=0.05, transform='log10')
deSp_plot <- plot_deSp_multiGroup(deSp_se)
}
