% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deSp_twoGroup.R
\name{plot_deSp_twoGroup}
\alias{plot_deSp_twoGroup}
\title{plot_deSp_twoGroup}
\usage{
plot_deSp_twoGroup(deSp_se)
}
\arguments{
\item{deSp_se}{A SummarizedExperiment object with results computed by \code{\link{deSp_twoGroup}}.}
}
\value{
Return a list of 3 interactive plots, 3 static plots, and 2 tables.
\enumerate{
\item interactive_de_lipid & static_de_lipid: a lollipop chart reveals the lipid species that pass chosen cut-off.
\item interactive_maPlot & static_maPlot: an MA plot of lipid species.
\item interactive_volcanoPlot & static_volcanoPlot: a volcano plot of lipid species.
\item table_de_lipid: table for plotting DE lipid plot.
\item table_ma_volcano: table for plotting MA plot and volcano plot.
}
}
\description{
This function is for plotting the results of lipid species differential expression analysis.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
deSp_se <- deSp_twoGroup(processed_se, ref_group='ctrl', test='t-test',
    significant='padj', p_cutoff=0.05, FC_cutoff=1, transform='log10')
deSp_plot <- plot_deSp_twoGroup(deSp_se)

}
