% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_enrichment_lsea.R
\name{plot_enrichment_lsea}
\alias{plot_enrichment_lsea}
\title{plot_enrichment_lsea}
\usage{
plot_enrichment_lsea(lsea_res, char, char_feature)
}
\arguments{
\item{lsea_res}{List. A result list of lipid-set enrichment analysis from \code{\link{enrichment_lsea}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}.}

\item{char_feature}{Character. A feature selected by users from \code{char}
to visualize the specific plot for the selected category of that characteristic.
For example, if char is 'class' and char_feature is 'Cer', the resulting
plots will display data for 'Cer' within the 'class' category.}
}
\value{
Return a list with 1 static plot and 1 interactive plot.
\enumerate{
\item static_enrichPlot: a static enriched plot of the selected lipid characteristic and feature.
\item interactive_enrichPlot: an interactive enriched plot of the selected lipid characteristic and feature.
}
}
\description{
This function plots the result of \code{\link{enrichment_lsea}}.
}
\examples{
data("de_data_twoGroup")
processed_se_twoGroup <- data_process(
    se=de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_list <- list_lipid_char(processed_se_twoGroup)$common_list
print(char_list)
deSp_se_twoGroup <- deSp_twoGroup(
    processed_se_twoGroup, ref_group='ctrl', test='t-test',
    significant='pval', p_cutoff=0.05, FC_cutoff=1, transform='log10')
lsea_one_twoGroup <- enrichment_lsea(
    deSp_se_twoGroup, char='class', rank_by='statistic', significant='pval',
    p_cutoff=0.05)
lsea_plot_twoGroup <- plot_enrichment_lsea(
    lsea_res=lsea_one_twoGroup, char='class', char_feature='TG')

data("se_multiGroup")
processed_se_multiGroup <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
char_list <- list_lipid_char(processed_se_multiGroup)$common_list
print(char_list)
deSp_se_multiGroup <- deSp_multiGroup(
    processed_se_multiGroup, ref_group='ctrl', test='One-way ANOVA',
    significant='pval', p_cutoff=0.05, transform='log10')
lsea_one_multiGroup <- enrichment_lsea(
    deSp_se_multiGroup, char='class', rank_by='statistic',
    significant='pval', p_cutoff=0.05)
lsea_plot_multiGroup <- plot_enrichment_lsea(
    lsea_res=lsea_one_multiGroup, char='class', char_feature='Cer')
}
