% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_feature.R
\name{plot_ml_feature}
\alias{plot_ml_feature}
\title{plot_ml_feature}
\usage{
plot_ml_feature(ml_se, feature_num = 10)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}
}
\value{
Return 2 interactive plots, 2 static plots, and 2 data frames.
\enumerate{
\item interactive_selected_frequency & static_selected_frequency: selected frequency plot
\item interactive_feature_importance & static_feature_importance: feature importance plot
\item table_selected_frequency: table for plotting selected frequency plot
\item table_feature_importance: table for plotting feature importance plot
}
}
\description{
This function computes and ranks each feature's contribution
based on the user-defined number of features and visualizes their importance.
When users select a specific number of features, the frequency of the top
features selected across all CV runs is displayed.
}
\examples{
data("ml_se_sub")
res <- plot_ml_feature(ml_se_sub, feature_num=10)
}
