% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_pr.R
\name{plot_ml_pr}
\alias{plot_ml_pr}
\title{plot_ml_pr}
\usage{
plot_ml_pr(ml_se, feature_num)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}
}
\value{
Return 2 interactive plots, 2 static plots, and 2 data frames.
\enumerate{
\item interactive_mean_auc & static_mean_auc: PR curve plots.
\item initeractive_pr & static_pr: PR Curve of average CVs plots.
\item table_mean_auc_plot: data frame of the AUC, recall, and precision of PR of n features.
\item table_pr: average PR curve plot of n feature.
}
}
\description{
This function generates the overall Precision-Recall (PR) curve
for cross-validations (CVs) with different feature numbers and the PR curve
for the average CVs based on user-selected feature numbers.
}
\examples{
data("ml_se_sub")
res <- plot_ml_pr(ml_se_sub, feature_num=10)
}
