% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_probability.R
\name{plot_ml_probability}
\alias{plot_ml_probability}
\title{plot_ml_probability}
\usage{
plot_ml_probability(ml_se, feature_num)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}
}
\value{
Return 1 interactive plots, 2 static plots, and 2 data frames.
\enumerate{
\item interactive_probability_plot & static_probability_plot: the distribution
of predicted probabilities in two reference labels.
\item static_confusion_matrix: A confusion matrix composed of sample number and proportion.
\item table_probability_plot: table for plotting probability plot.
\item table_confusion_matrix: table for plotting confusion matrix.
}
}
\description{
This function computes and visualizes the average predicted
probabilities for each sample in the testing data across all cross-validation
(CV) runs. The function plots the distribution of predicted probabilities for
two reference labels and a confusion matrix that includes both the sample count
and proportion.
}
\examples{
data("ml_se_sub")
res <- plot_ml_probability(ml_se_sub, feature_num=10)

}
