% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ml_roc.R
\name{plot_ml_roc}
\alias{plot_ml_roc}
\title{plot_ml_roc}
\usage{
plot_ml_roc(ml_se, feature_num = 10)
}
\arguments{
\item{ml_se}{A SummarizedExperiment object with results computed by \code{\link{ml_model}}.}

\item{feature_num}{Numeric. The number of features to be shown in the plots.
A feature number value selected from \code{feature_option} of the
SummarizedExperiment object returned by \code{\link{ml_model}}. Usually be
one of 2, 3, 5, 10, 20, 50, 100. Default is \code{10}.}
}
\value{
Return 2 interactive plots, 2 static plots, and 2 data frames.
\enumerate{
\item interactive_mean_auc & static_mean_auc: ROC curve plots
\item initeractive_roc & static_roc: ROC Curve of average CVs plots
\item table_mean_auc_plot: ROC data frame of n features
\item table_roc: average ROC curve plot of n features
}
}
\description{
This function generates the overall ROC curve for cross-validations
(CVs) with varying feature numbers and the ROC curve for the average CVs based
on user-selected feature numbers.
}
\examples{
data("ml_se_sub")
res <- plot_ml_roc(ml_se_sub, feature_num=10)
}
