% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{plot_ml_shap}
\alias{plot_ml_shap}
\title{plot_ml_shap}
\usage{
plot_ml_shap(shap_se)
}
\arguments{
\item{shap_se}{A SummarizedExperiment object with results computed by \code{\link{ml_shap}}.}
}
\value{
Return 2 interactive plots, 2 static plots, and 2 tables.
\enumerate{
\item interactive_feature_importance & static_feature_importance: SHAP feature importance plot.
\item interactive_summary_plot & static_summary_plot: SHAP value plot.
\item table_feature_importance: table for plotting feature importance plot.
\item table_summary_plot: table for plotting SHAP value plot.
}
}
\description{
This function is for visualizing the results of Shapley Additive
exPlanations (SHAP) results.
}
\examples{
library(SHAPforxgboost)
data("ml_se_sub")
shap_se <- ml_shap(ml_se_sub, feature_num=10, nsim=5)
res <- plot_ml_shap(shap_se)
}
