% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{plot_shap_dependence}
\alias{plot_shap_dependence}
\title{plot_shap_dependence}
\usage{
plot_shap_dependence(shap_se, feature, shap_feature, interaction_index)
}
\arguments{
\item{shap_se}{A SummarizedExperiment object with results computed by \code{\link{ml_shap}}.}

\item{feature}{Character. A character string of the feature name the user selects
from the "variable" column in \code{shap_result}. (Use \code{\link{extract_summarized_experiment}}
to view \code{shap_result}.)}

\item{shap_feature}{Character. A character string of the feature name the user selects from the
"variable" column of \code{shap_result}. (Use \code{\link{extract_summarized_experiment}}
to view \code{shap_result}.)}

\item{interaction_index}{Character. A character string of the feature name the user selects
from the "variable" column of \code{shap_result}. (Use \code{\link{extract_summarized_experiment}}
to view \code{shap_result}.)}
}
\value{
Return 1 interactive plot, 1 static plot, 1 table.
\enumerate{
\item interactive_dependence_plot ＆ static_dependence_plot: SHAP dependence plot.
\item table_dependence_plot: table for plotting SHAP dependence plot
}
}
\description{
This function visualizes the SHAP values against the feature
values for each variable.
}
\examples{
data("ml_se_sub")
shap_se <- shap(ml_se_sub, feature_num=10, nsim=5)
selected_feature <- as.character(unique(S4Vectors::metadata(shap_se)$shap_result$variable))
res <- plot_shap_dependence(
    shap_se, feature=selected_feature[1], shap_feature=selected_feature[2],
    interaction_index=selected_feature[2])
}
