% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml_shap.R
\name{plot_shap_force}
\alias{plot_shap_force}
\title{plot_shap_force}
\usage{
plot_shap_force(
  shap_se,
  top_feature = 10,
  cluster_method = "ward.D",
  group_num = 10
)
}
\arguments{
\item{shap_se}{A SummarizedExperiment object with results computed by \code{\link{ml_shap}}.}

\item{top_feature}{Integer. Top number of features to be shown. If top_feature
is set to greater than 10, only the top 10 samples will be displayed. Default is \code{10}.}

\item{cluster_method}{Character. The clustering method. Allowed methods include
"ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid".
Default is \code{'ward.D'}.}

\item{group_num}{Integer. The numbers of groups to be shown in the plots.
Default is \code{10}.}
}
\value{
Return 1 interactive plot, 1 static plot, and 1 table.
\enumerate{
\item interactive_forcePlot & static_forcePlot: SHAP force plot.
\item table_forcePlot: table for plotting force plot.
}
}
\description{
The function stacks the SHAP values for each observation and shows
how the final output was obtained as a sum of each predictor’s attributions through
the force plot. (Randomly plotting a certain portion of the data is optional in
case the dataset is large.)
}
\examples{
data("ml_se_sub")
shap_se <- ml_shap(ml_se_sub, feature_num=10, nsim=5)
res <- plot_shap_force(shap_se, top_feature=10, cluster_method="ward.D", group_num=10)
}
