% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_subChar_multiGroup.R
\name{plot_subChar_multiGroup}
\alias{plot_subChar_multiGroup}
\title{plot_subChar_multiGroup}
\usage{
plot_subChar_multiGroup(subChar_se, subChar_feature)
}
\arguments{
\item{subChar_se}{A SummarizedExperiment object with results computed by \code{\link{subChar_multiGroup}}.}

\item{subChar_feature}{Character. A feature selected by users from \code{subChar} to visualize the specific plot for the selected category of that characteristic.
For example, if subChar is 'class' and subChar_feature is 'Cer', the resulting plots will display data for 'Cer' within the 'class' category.}
}
\value{
Return a list with 5 static plots, 4 interactive plots, and 5 data frames.
\enumerate{
\item static_barPlot: a static bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic.
\item static_barPlot_sqrt: a static bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_linePlot: a static line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic.
\item static_linePlot_sqrt: a static line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_boxPlot: a static box plot of ctrl group and experiment group.
\item interactive_barPlot: an interactive bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic.
\item interactive_barPlot_sqrt: an interactive bar plot shows the average expression of each sample group and highlights significant differences between groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item interactive_linePlot: an interactive line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic.
\item interactive_linePlot_sqrt: an interactive line plot shows the average expression of each sample group and highlights significant differences between the groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item table_barPlot: table for plotting bar plots
\item table_linePlot: table for plotting line plots
\item table_boxPlot: table for plotting box plots
\item table_char_index: table with the value calculated by the weighted average of lipid characteristics abundance
\item table_index_stat: table with statistics of control and experiment groups
}
}
\description{
This function is for plotting the results of subgroup lipid characteristics differential expression analysis.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
subChar_se <- subChar_multiGroup(
    processed_se, char='Total.C', subChar='class', ref_group='ctrl',
    post_hoc='One-way ANOVA', post_hoc_sig='pval', post_hoc_p_cutoff=0.05,
    transform='log10')
subChar_plot <- plot_subChar_multiGroup(subChar_se, subChar_feature='CL')
}
