% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subChar_twoGroup.R
\name{plot_subChar_twoGroup}
\alias{plot_subChar_twoGroup}
\title{plot_subChar_twoGroup}
\usage{
plot_subChar_twoGroup(subChar_se, subChar_feature)
}
\arguments{
\item{subChar_se}{A SummarizedExperiment object with results computed by \code{\link{subChar_twoGroup}}.}

\item{subChar_feature}{Character. A feature selected by users from \code{subChar} to visualize the specific plot for the selected category of that characteristic.
For example, if subChar is 'class' and subChar_feature is 'Cer', the resulting plots will display data for 'Cer' within the 'class' category.}
}
\value{
Return a list with 5 static plots, 5 interactive plots, and 5 data frames.
\enumerate{
\item static_barPlot: a static bar plot shows the average expression of each sample group and highlights significant
differences between two groups based on a user-selected characteristic.
\item static_barPlot_sqrt: a static bar plot shows the average expression of each sample
group and highlights significant differences between two groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_linePlot: a static line plot shows the average expression of each sample group and
highlights significant differences between the two groups based on a user-selected characteristic.
\item static_linePlot_sqrt: a static line plot shows the average expression of each sample group and highlights
significant differences between the two groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item static_boxPlot: a static box plot of ctrl group and experiment group.
\item interactive_barPlot: an interactive bar plot shows the average expression of each sample group
and highlights significant differences between two groups based on a user-selected characteristic.
\item interactive_barPlot_sqrt: an interactive bar plot shows the average expression of each sample
group and highlights significant differences between two groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item interactive_linePlot: an interactive line plot shows the average expression of each
sample group and highlights significant differences between the two groups based on a user-selected characteristic.
\item interactive_linePlot_sqrt: an interactive line plot shows the average expression of
each sample group and highlights significant differences between the two groups based on a user-selected characteristic. \emph{NOTE: the y axis is sqrt-scaled}.
\item interactive_boxPlot: an interactive box plot of ctrl group and experiment group.
\item table_barPlot: table for plotting bar plots
\item table_linePlot: table for plotting line plots
\item table_boxPlot: table for plotting box plots
\item table_char_index: table with the value calculated by the weighted average of lipid characteristics abundance
\item table_index_stat: table with statistics of control and experiment groups
}
}
\description{
This function is for plotting the results of subgroup lipid characteristics differential expression analysis.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage', transform='log10')
subChar_se <- subChar_twoGroup(processed_se, char="Total.C", subChar="class",
    ref_group="ctrl", test='t-test', significant="pval", p_cutoff=0.05,
    FC_cutoff=1, transform='log10')
subChar_feature_list <- unique(
    extract_summarized_experiment(subChar_se)$all_deChar_result$sub_feature)
res_plot <- plot_subChar_twoGroup(subChar_se, subChar_feature="Cer")
}
