% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLSDA.R
\name{PLSDA}
\alias{PLSDA}
\title{PLSDA}
\usage{
PLSDA(
  exp_transform_table,
  group_info = NULL,
  sig_feature = NULL,
  ncomp = 2,
  scaling = TRUE,
  cluster_method,
  group_num = NULL,
  var1 = NULL,
  var2 = NULL,
  insert_ref_group = NULL,
  ref_group = NULL,
  plotly = TRUE
)
}
\arguments{
\item{exp_transform_table}{A data frame of predictors, including
features (molecule, lipid class, etc.) and their expression of each sample.
NAs are not allowed. The name of the first column must be
"feature" (lipid species).}

\item{group_info}{A data frame with sample name, label name, the group name
of samples (ctrl & exp), and pair number representing 'the pair' for the
t-test/Wilcoxon test. NAs are allowed.}

\item{sig_feature}{A character string for identifying significant variable
in \bold{exp_transform_table}.}

\item{ncomp}{A numeric value. The number of components to include in
the model. (default: 2)}

\item{scaling}{A logical value. If scaling = TRUE, each block is
standardized to zero means and unit variances. (default: TRUE).}

\item{cluster_method}{A character string indicating which method to be used
for clustering. Allowed method include \bold{"kmeans"}, \bold{"kmedoids"},
\bold{"hclustering"}, \bold{"dbscan"}, \bold{"group_info"}.}

\item{group_num}{A positive integer specifying the number of clusters.
Group number must be between 1 and 10.}

\item{var1}{A character string or numeric. According to the
selected cluster method.
\itemize{
  \item \emph{kmeans}: A specifying the metric to be used for calculating
  dissimilarities between observations include \bold{"euclidean"}
  and \bold{"manhattan"}.
  \item \emph{hclustering}: which correlation coefficient is to be computed
  or which the distance measure to be used include \bold{"pearson"},
  \bold{"kendall"}, \bold{"spearman"}, \bold{"euclidean"},
  \bold{"manhattan"}, \bold{"maximum"}, \bold{"canberra"},
  \bold{"binary"}, \bold{"minkowski"}.
  \item \emph{dbscan}: size of the epsilon neighborhood.
}}

\item{var2}{A character string or numeric. According to the
selected cluster method.
\itemize{
  \item \emph{hclustering}: The agglomeration method to be used.
  This should be (an unambiguous abbreviation of) one of \bold{"ward.D"},
  \bold{"ward.D2"}, \bold{"single"}, \bold{"complete"},
  \bold{"average"}(= UPGMA), \bold{"mcquitty"}(= WPGMA),
  \bold{"median"}(= WPGMC), or \bold{"centroid"}(= UPGMC).
  \item \emph{dbscan}: number of minimum points in the
  eps region (for core points).
}}

\item{insert_ref_group}{A character string. The name of 'ctrl' after
name conversion.}

\item{ref_group}{A character string. The name of 'exp' after
name conversion.}

\item{plotly}{Logical value. If TRUE, return the resulting plots dynamically.}
}
\value{
Return a list with 2 data frames and 2 plots.
\enumerate{
\item X.variate: a data frame of sample variate associated to each
component if PLS-DA is applied. This is for \bold{sample.plot}.
\item X.loading: a data frame of sample loading, which are coefficients
assigned to each variable to define each component. These coefficients
indicate the importance of each variable in PLS-DA. This plot is for
\bold{variable.plot}.
\item sample.plot: PLS-DA sample plot. The axis labels indicate the
amount of variation explained per component.
\item variable.plot: PLS-DA variable plot, display the variables that
contribute to the definition of each component.  These variables should be
located towards the circle of radius 1, far from the center.
}
}
\description{
A method of dimensionality reduction to transform data from a
high-dimensional space into a low-dimensional space so that the vital
properties of the original data can be retained.
}
\examples{
data("DE_exp_data")
data("DE_group_info")
exp_data <- DE_exp_data
group_info <- DE_group_info
exp_transform_table <- data_process(exp_data, exclude_var_missing=TRUE,
                                    missing_pct_limit=50, replace_zero=TRUE,
                                    zero2what='min', xmin=0.5,
                                    replace_NA=TRUE, NA2what='min',
                                    ymin=0.5,  pct_transform=TRUE,
                                    data_transform=FALSE, trans_type='log',
                                    centering=FALSE, scaling=FALSE)
DE_species_table_sig <- DE_species_2(exp_transform_table,
                                     data_transform=TRUE,
                                     group_info=group_info, paired=FALSE,
                                     test='t.test', adjust_p_method='BH',
                                     sig_stat='p.adj', sig_pvalue=0.05,
                                     sig_FC=2,
                                     plotly=TRUE)$DE_species_table_sig
PLSDA(exp_transform_table, group_info=group_info,
      sig_feature=DE_species_table_sig$feature,
      ncomp=2, scaling=TRUE, cluster_method='group_info',
      group_num=NULL, var1=NULL, var2=NULL,
      insert_ref_group=NULL, ref_group=NULL, plotly=TRUE)
}
