\name{profiling_lipid_char_table}
\alias{profiling_lipid_char_table}
\docType{data}
\title{
profiling_lipid_char_table
}
\description{
A data frame of lipid characteristics for profiling analysis workflow.
}
\usage{data("profiling_lipid_char_table")}
\format{
  A data frame with 208 observations on the following 10 variables.
  \describe{
    \item{\code{feature}}{a character vector}
    \item{\code{class}}{a character vector}
    \item{\code{structural_category}}{a character vector}
    \item{\code{functional_category}}{a character vector}
    \item{\code{totallength}}{a numeric vector}
    \item{\code{totaldb}}{a numeric vector}
    \item{\code{totaloh}}{a numeric vector}
    \item{\code{FA_length}}{a numeric vector}
    \item{\code{FA_db}}{a character vector}
    \item{\code{FA_oh}}{a character vector}
  }
}
\examples{
data(profiling_lipid_char_table)
}
\keyword{datasets}
