% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Species2Char.R
\name{Species2Char}
\alias{Species2Char}
\title{Species2Char}
\usage{
Species2Char(exp_data, lipid_char_table, char_var)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}

\item{lipid_char_table}{A data frame of lipid characteristics such as name(feature) of lipid, class of lipid, the total length of lipid, and Fatty acid (FA_) related characteristics. NAs are allowed. The name of the first column must be "feature" (lipid species).}

\item{char_var}{A character string of the lipid characteristic selected by users from the column name of \bold{lipid_char_table}, such as 'class'}
}
\value{
Return \bold{transform_table}. An data frame of aggregated(sum) expression data by user-selected characteristics (\bold{char_var}).
}
\description{
Summary the expression in \bold{exp_data} from lipid species to user-selected characteristics.
}
\examples{
\donttest{
data("DE_exp_data")
exp_data <- DE_exp_data
lipid_char_table <- data("DE_lipid_char_table")
char_var <- colnames(lipid_char_table)[-1]
Species2Char(exp_data, lipid_char_table, char_var=char_var[1])
}
}
