% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subChar_multiGroup.R
\name{subChar_multiGroup}
\alias{subChar_multiGroup}
\title{subChar_multiGroup}
\usage{
subChar_multiGroup(
  processed_se,
  char,
  subChar,
  ref_group,
  post_hoc = c("One-way ANOVA", "Kruskal–Wallis test"),
  post_hoc_sig = c("pval", "padj"),
  post_hoc_p_cutoff = 0.05,
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}. Allowed lipid characteristics include
'Total.C', 'Total.DB', 'Total.OH', 'FA.C', 'FA.DB', and 'FA.OH'.
Selected characteristics must be one of 'Total.C', 'Total.DB', 'Total.OH', 'FA.C', 'FA.DB', or 'FA.OH'.}

\item{subChar}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}. Must be differ from 'char'.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{post_hoc}{Character. The method to statistic calculation. Allowed method
include "One-way ANOVA" and "Kruskal–Wallis test". Default is \code{'One-way ANOVA'}.}

\item{post_hoc_sig}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{post_hoc_p_cutoff}{Numeric. Significant level. Default is \code{0.05}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function conducts subgroup lipid characteristic analysis for multi-group data.
Lipid species are categorized and summarized into a new lipid abundance table
according to two selected lipid characteristics, followed by differential expression analysis.
The two chosen characteristics should be either both continuous data or one continuous and one categorical data.
}
\examples{
data("se_multiGroup")
processed_se <- data_process(
    se=se_multiGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
subChar_se <- subChar_multiGroup(
    processed_se, char='Total.C', subChar='class', ref_group='ctrl',
    post_hoc='One-way ANOVA', post_hoc_sig='pval', post_hoc_p_cutoff=0.05,
    transform='log10')
}
