% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subChar_twoGroup.R
\name{subChar_twoGroup}
\alias{subChar_twoGroup}
\title{subChar_twoGroup}
\usage{
subChar_twoGroup(
  processed_se,
  char,
  subChar,
  ref_group,
  test = c("t-test", "Wilcoxon test"),
  significant = c("pval", "padj"),
  p_cutoff = 0.05,
  FC_cutoff = 1,
  transform = c("none", "log10", "square", "cube")
)
}
\arguments{
\item{processed_se}{A SummarizedExperiment object constructed by
\code{\link{as_summarized_experiment}} and processed by \code{\link{data_process}}.}

\item{char}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}.}

\item{subChar}{Character. A lipid characteristic selected from the common list
returned by \code{\link{list_lipid_char}}. Must be differ from 'char'.}

\item{ref_group}{Character. Group name of the reference group. It must be one of the
group names in the group information table's group column.}

\item{test}{A character string indicating which method to be used for
comparing means. Allowed method include \bold{"t-test"} and
\bold{"Wilcoxon test"}. Default is \code{'t-test'}.}

\item{significant}{Character. The p-value to be used for the statistically
significant. Must be one of "pval" or "padj". Default is \code{'pval'}.}

\item{p_cutoff}{Numeric. Significant level. Default is \code{0.05}.}

\item{FC_cutoff}{Numeric. Significance of the fold-change. Default is \code{1}.}

\item{transform}{Character. Method for data transformation. Allowed methods
include "none", "log10", "square", "cube". Select 'none' to skip data transformation.
Default is \code{'log10'}.}
}
\value{
Return a SummarizedExperiment object containing analysis results.
}
\description{
This function conducts subgroup lipid characteristic analysis for two-group data.
Lipid species are categorized and summarized into a new lipid abundance table
according to two selected lipid characteristics, followed by differential expression analysis.
The two chosen characteristics should be either both continuous data or one continuous and one categorical data.
}
\examples{
data("de_data_twoGroup")
processed_se <- data_process(
    de_data_twoGroup, exclude_missing=TRUE, exclude_missing_pct=70,
    replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
subChar_se <- subChar_twoGroup(processed_se, char="Total.C", subChar="class",
    ref_group="ctrl", test='t-test', significant="pval", p_cutoff=0.05,
    FC_cutoff=1, transform='log10')
}
