% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UMAP.R
\name{UMAP}
\alias{UMAP}
\title{UMAP}
\usage{
UMAP(
  exp_transform_table,
  group_info = NULL,
  sig_feature = NULL,
  n_neighbors = 15,
  scale = T,
  metric = "euclidean",
  cluster_method,
  group_num = NULL,
  var1 = NULL,
  var2 = NULL,
  insert_ref_group = NULL,
  ref_group = NULL
)
}
\arguments{
\item{exp_transform_table}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}

\item{group_info}{A data frame for marking clustering sample on PCA chart. It is according to \bold{Group information table}(after convert to ctrl and exp).}

\item{sig_feature}{A character string for identifying significant variable in \bold{exp_transform_table}.}

\item{n_neighbors}{The size of local neighborhood (in terms of number of neighboring sample points) used for manifold approximation.}

\item{scale}{Scaling to apply to \bold{X} if it is a data frame or matrix:
\itemize{
  \item \bold{"none"} or \bold{FALSE} or \bold{NULL}: No scaling.
  \item \bold{"Z"} or \bold{"scale"} or \bold{TRUE}: Scale each column to zero mean and variance 1.
  \item \bold{"maxabs"}: Center each column to mean 0, then divide each element by the maximum absolute value over the entire matrix.
  \item \bold{"range"}: Range scale the entire matrix, so the smallest element is 0 and the largest is 1.
  \item \bold{"colrange"}: Scale each column in the range (0,1).
}}

\item{metric}{Type of distance metric to use to find nearest neighbors. One of \bold{"euclidean"}, \bold{"cosine"}, \bold{"manhattan"}, \bold{"hamming"}, \bold{"categorical"}.}

\item{cluster_method}{A character string indicating which method to be used for clustering. Allowed method include \bold{"kmeans"}, \bold{"kmedoids"}, \bold{"hclustering"}, \bold{"dbscan"}, \bold{"group_info"}.}

\item{group_num}{A positive integer specifying the number of clusters. Group number must be between 1 and 10'}

\item{var1}{A character string or numeric. According to the selected cluster method.
\itemize{
  \item \emph{kmeans}: specifying the metric to be used for calculating dissimilarities between observations include \bold{"euclidean"} and \bold{"manhattan"}.
  \item \emph{hclustering}: which correlation coefficient is to be computed or which the distance measure to be used include \bold{"pearson"}, \bold{"kendall"}, \bold{"spearman"}, \bold{"euclidean"}, \bold{"manhattan"}, \bold{"maximum"}, \bold{"canberra"}, \bold{"binary"}, \bold{"minkowski"}.
  \item \emph{dbscan}: size of the epsilon neighborhood.
}}

\item{var2}{A character string or numeric. According to the selected cluster method.
\itemize{
  \item \emph{hclustering}: The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of \bold{"ward.D"}, \bold{"ward.D2"}, \bold{"single"}, \bold{"complete"}, \bold{"average"}(=UPGMA), \bold{"mcquitty"}(= WPGMA), \bold{"median"}(= WPGMC), or \bold{"centroid"} (= UPGMC).
  \item \emph{dbscan}: number of minimum points in the eps region (for core points).
}}

\item{insert_ref_group}{A character string. The name of 'ctrl' after name conversion.}

\item{ref_group}{A character string. The name of 'exp' after name conversion.}
}
\value{
Return a list with 1 data frame and 1 plot.
\enumerate{
\item umap_data: a data frame of UMAP data.
\item umap_plot: UMAP plot
}
}
\description{
A method of dimensionality reduction to transform data from a high-dimensional space into a low-dimensional space so that the vital properties of the original data can be retained.
}
\examples{
\dontrun{
data("profiling_exp_data")
exp_data <- profiling_exp_data
exp_transform_table <- data_process(exp_data, exclude_var_missing=T, missing_pct_limit=50,
                                    replace_zero=T, zero2what='min', xmin=0.5, replace_NA=T,
                                    NA2what='min', ymin=0.5, pct_transform=T, data_transform=T,
                                    trans_type='log', centering=F,  scaling=F)
UMAP(exp_transform_table, group_info = NULL, sig_feature = NULL,n_neighbors=15, scale=T,
     metric='euclidean', cluster_method='kmeans',group_num = 2, var1 = NULL, var2 = NULL,
     insert_ref_group = NULL, ref_group = NULL)
}
}
