\name{corr_exp_data}
\alias{corr_exp_data}
\docType{data}
\title{
corr_exp_data
}
\description{
A data frame of lipid expression data for correlation analysis workflow. The
lipid expression data is obtained from the cohort interrogating the associations
of plasma sphingolipids with subphenotypes of COPD, including airflow
obstruction, emphysema, and frequent exacerbations. The dataset contained 129
current and former smokers.
}
\usage{data("corr_exp_data")}
\format{
  A data frame with 69 observations on the following 130 variables.
  \describe{
    \item{\code{feature}}{a character vector}
    \item{\code{sample1}}{a numeric vector}
    \item{\code{sample2}}{a numeric vector}
    \item{\code{sample3}}{a numeric vector}
    \item{\code{sample4}}{a numeric vector}
    \item{\code{sample5}}{a numeric vector}
    \item{\code{sample6}}{a numeric vector}
    \item{\code{sample7}}{a numeric vector}
    \item{\code{sample8}}{a numeric vector}
    \item{\code{sample9}}{a numeric vector}
    \item{\code{sample10}}{a numeric vector}
    \item{\code{sample11}}{a numeric vector}
    \item{\code{sample12}}{a numeric vector}
    \item{\code{sample13}}{a numeric vector}
    \item{\code{sample14}}{a numeric vector}
    \item{\code{sample15}}{a numeric vector}
    \item{\code{sample16}}{a numeric vector}
    \item{\code{sample17}}{a numeric vector}
    \item{\code{sample18}}{a numeric vector}
    \item{\code{sample19}}{a numeric vector}
    \item{\code{sample20}}{a numeric vector}
    \item{\code{sample21}}{a numeric vector}
    \item{\code{sample22}}{a numeric vector}
    \item{\code{sample23}}{a numeric vector}
    \item{\code{sample24}}{a numeric vector}
    \item{\code{sample25}}{a numeric vector}
    \item{\code{sample26}}{a numeric vector}
    \item{\code{sample27}}{a numeric vector}
    \item{\code{sample28}}{a numeric vector}
    \item{\code{sample29}}{a numeric vector}
    \item{\code{sample30}}{a numeric vector}
    \item{\code{sample31}}{a numeric vector}
    \item{\code{sample32}}{a numeric vector}
    \item{\code{sample33}}{a numeric vector}
    \item{\code{sample34}}{a numeric vector}
    \item{\code{sample35}}{a numeric vector}
    \item{\code{sample36}}{a numeric vector}
    \item{\code{sample37}}{a numeric vector}
    \item{\code{sample38}}{a numeric vector}
    \item{\code{sample39}}{a numeric vector}
    \item{\code{sample40}}{a numeric vector}
    \item{\code{sample41}}{a numeric vector}
    \item{\code{sample42}}{a numeric vector}
    \item{\code{sample43}}{a numeric vector}
    \item{\code{sample44}}{a numeric vector}
    \item{\code{sample45}}{a numeric vector}
    \item{\code{sample46}}{a numeric vector}
    \item{\code{sample47}}{a numeric vector}
    \item{\code{sample48}}{a numeric vector}
    \item{\code{sample49}}{a numeric vector}
    \item{\code{sample50}}{a numeric vector}
    \item{\code{sample51}}{a numeric vector}
    \item{\code{sample52}}{a numeric vector}
    \item{\code{sample53}}{a numeric vector}
    \item{\code{sample54}}{a numeric vector}
    \item{\code{sample55}}{a numeric vector}
    \item{\code{sample56}}{a numeric vector}
    \item{\code{sample57}}{a numeric vector}
    \item{\code{sample58}}{a numeric vector}
    \item{\code{sample59}}{a numeric vector}
    \item{\code{sample60}}{a numeric vector}
    \item{\code{sample61}}{a numeric vector}
    \item{\code{sample62}}{a numeric vector}
    \item{\code{sample63}}{a numeric vector}
    \item{\code{sample64}}{a numeric vector}
    \item{\code{sample65}}{a numeric vector}
    \item{\code{sample66}}{a numeric vector}
    \item{\code{sample67}}{a numeric vector}
    \item{\code{sample68}}{a numeric vector}
    \item{\code{sample69}}{a numeric vector}
    \item{\code{sample70}}{a numeric vector}
    \item{\code{sample71}}{a numeric vector}
    \item{\code{sample72}}{a numeric vector}
    \item{\code{sample73}}{a numeric vector}
    \item{\code{sample74}}{a numeric vector}
    \item{\code{sample75}}{a numeric vector}
    \item{\code{sample76}}{a numeric vector}
    \item{\code{sample77}}{a numeric vector}
    \item{\code{sample78}}{a numeric vector}
    \item{\code{sample79}}{a numeric vector}
    \item{\code{sample80}}{a numeric vector}
    \item{\code{sample81}}{a numeric vector}
    \item{\code{sample82}}{a numeric vector}
    \item{\code{sample83}}{a numeric vector}
    \item{\code{sample84}}{a numeric vector}
    \item{\code{sample85}}{a numeric vector}
    \item{\code{sample86}}{a numeric vector}
    \item{\code{sample87}}{a numeric vector}
    \item{\code{sample88}}{a numeric vector}
    \item{\code{sample89}}{a numeric vector}
    \item{\code{sample90}}{a numeric vector}
    \item{\code{sample91}}{a numeric vector}
    \item{\code{sample92}}{a numeric vector}
    \item{\code{sample93}}{a numeric vector}
    \item{\code{sample94}}{a numeric vector}
    \item{\code{sample95}}{a numeric vector}
    \item{\code{sample96}}{a numeric vector}
    \item{\code{sample97}}{a numeric vector}
    \item{\code{sample98}}{a numeric vector}
    \item{\code{sample99}}{a numeric vector}
    \item{\code{sample100}}{a numeric vector}
    \item{\code{sample101}}{a numeric vector}
    \item{\code{sample102}}{a numeric vector}
    \item{\code{sample103}}{a numeric vector}
    \item{\code{sample104}}{a numeric vector}
    \item{\code{sample105}}{a numeric vector}
    \item{\code{sample106}}{a numeric vector}
    \item{\code{sample107}}{a numeric vector}
    \item{\code{sample108}}{a numeric vector}
    \item{\code{sample109}}{a numeric vector}
    \item{\code{sample110}}{a numeric vector}
    \item{\code{sample111}}{a numeric vector}
    \item{\code{sample112}}{a numeric vector}
    \item{\code{sample113}}{a numeric vector}
    \item{\code{sample114}}{a numeric vector}
    \item{\code{sample115}}{a numeric vector}
    \item{\code{sample116}}{a numeric vector}
    \item{\code{sample117}}{a numeric vector}
    \item{\code{sample118}}{a numeric vector}
    \item{\code{sample119}}{a numeric vector}
    \item{\code{sample120}}{a numeric vector}
    \item{\code{sample121}}{a numeric vector}
    \item{\code{sample122}}{a numeric vector}
    \item{\code{sample123}}{a numeric vector}
    \item{\code{sample124}}{a numeric vector}
    \item{\code{sample125}}{a numeric vector}
    \item{\code{sample126}}{a numeric vector}
    \item{\code{sample127}}{a numeric vector}
    \item{\code{sample128}}{a numeric vector}
    \item{\code{sample129}}{a numeric vector}
  }
}
\source{
Bowler, Russell P., et al. "Plasma sphingolipids associated with chronic
obstructive pulmonary disease phenotypes." American journal of respiratory
and critical care medicine 191.3(2015):275-284.
}
\examples{
data(corr_exp_data)
}
\keyword{datasets}
