% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{corr_heatmap}
\usage{
corr_heatmap(
  exp_data,
  corr_method = "pearson",
  distfun = "maximum",
  hclustfun = "average"
)
}
\arguments{
\item{exp_data}{A data frame of predictors, including features (molecule, lipid class, etc.) and their expression of each sample. NAs are not allowed. The name of the first column must be "feature" (lipid species).}

\item{corr_method}{A character string indicating which correlation coefficient is to be computed. One of "pearson" or "spearman". (default: "pearson")}

\item{distfun}{A character string of the distance measure indicating which correlation coefficient (or covariance) is to be computed. Allowed methods include "pearson", "spearman", "kendall", euclidean", "maximum", "manhattan", "canberra", "binary", "minkowski". (default: "maximum")}

\item{hclustfun}{A character string of the agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). (default: "average")}
}
\value{
Return 2 plots and a list of 6 matrices.
\enumerate{
\item sample_corr_coef: a matrix of correlation coefficients between samples
\item sample_corr_p: a matrix of correlation p-value between samples
\item reorder_sample_corr_coef: a matrix of reorder sample correlation.
\item lipids_corr_coef: a matrix of correlation coefficients between features (lipids).
\item lipids_corr_p: a matrix of correlation p-value between features (lipids).
\item reorder_lipids_corr_coef: reorder features (lipids) correlation coefficients matrix
\item heatmap plot by samples
\item heatmap plot by lipid species
}
}
\description{
The correlation heatmap illustrates the correlation between lipid species or samples and depicts the patterns in each group. The correlation can be calculated by Pearson or Spearman.
\enumerate{
\item Compute and cluster the correlation coefficient depending on user-defined method and distance.
\item Print two correlation heatmaps to illustrate the correlation between lipid species or samples.
}
}
\examples{
\donttest{
data("profiling_exp_data")
exp_data <- profiling_exp_data
exp_transform <- data_process(exp_data, exclude_var_missing=T, missing_pct_limit=50,
                              replace_zero=T, zero2what='min', xmin=0.5, replace_NA=T,
                              NA2what='min', ymin=0.5, pct_transform=T, data_transform=T,
                              trans_type='log', centering=F,  scaling=F)
corr_heatmap(exp_transform, corr_method="pearson", distfun="maximum", hclustfun="average")
}
}
