% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_heatmap.R
\name{corr_heatmap}
\alias{corr_heatmap}
\title{corr_heatmap}
\usage{
corr_heatmap(
  exp_data_SE,
  corr_method = "pearson",
  distfun = "maximum",
  hclustfun = "average",
  type = "sample"
)
}
\arguments{
\item{exp_data_SE}{A SummarizedExperiment object contains information about
various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
allowed. The row data  corresponds to specific lipid features, such as class
and total length. The first column's name must be "feature" (lipid species),
and NAs are allowed for this data. The column data comprises sample names,
sample labels, group names, and pair numbers that represent 'the pair' for
conducting t-tests or Wilcoxon tests. NAs are allowed.}

\item{corr_method}{A character string indicating which correlation
coefficient is to be computed. One of "pearson" or
"spearman". (default: "pearson")}

\item{distfun}{A character string of the distance measure indicating which
correlation coefficient (or covariance) is to be computed. Allowed methods
include "pearson", "spearman", "kendall", euclidean", "maximum",
"manhattan", "canberra", "binary", "minkowski". (default: "maximum")}

\item{hclustfun}{A character string of the agglomeration method to be used.
This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2",
"single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA),
"median" (= WPGMC) or "centroid" (= UPGMC). (default: "average")}

\item{type}{It should be 'sample' or 'lipid.' 'sample' outputs the
correlation results of samples, and 'lipid' outputs output the correlation
results of lipid species.}
}
\value{
Return a SummarizedExperiment object.
}
\description{
The correlation heatmap illustrates the correlation between
lipid species or samples and depicts the patterns in each group.
The correlation can be calculated by Pearson or Spearman.
\enumerate{
\item Compute and cluster the correlation coefficient depending on
user-defined method and distance.
\item Print two correlation heatmaps to illustrate the correlation between
lipid species or samples.
}
}
\examples{
data("profiling_data")
exp_data_SE <- profiling_data
exp_transform_SE <- data_process(exp_data_SE, exclude_var_missing=TRUE,
    missing_pct_limit=50, replace_zero=TRUE, zero2what='min', xmin=0.5,
    replace_NA=TRUE, NA2what='min', ymin=0.5, pct_transform=TRUE,
    data_transform=TRUE, trans_type='log', centering=FALSE, scaling=FALSE)
corr_heatmap(exp_transform_SE, corr_method="pearson", distfun="maximum",
    hclustfun="average",type='sample')
}
