% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{data_process}
\usage{
data_process(
  se,
  exclude_missing = TRUE,
  exclude_missing_pct = 70,
  replace_na_method = c("none", "QRILC", "SVD", "KNN", "IRMI", "min", "mean", "median",
    "PPCA", "BPCA", "RandomForest"),
  replace_na_method_ref = 0.5,
  normalization = c("none", "Percentage", "PQN", "Quantile", "Sum", "Median")
)
}
\arguments{
\item{se}{A SummarizedExperiment object construct by \code{\link{as_summarized_experiment}}.}

\item{exclude_missing}{Logical. If exclude_missing=TURE, lipids with
missing values will be removed. Default is \code{TRUE}.}

\item{exclude_missing_pct}{Numeric. Lipids with missing values over a
certain percentage (5-100) should be removed. Default is \code{70}.}

\item{replace_na_method}{Character. The method for NA values replacing.
Allowed methods include "QRILC", "SVD", "KNN", "IRMI", "min", "mean", "median",
"PPCA", "BPCA", "RandomForest", and "none". If you have already replaced NAs,
select 'none'. Default is \code{'min'}.}

\item{replace_na_method_ref}{Numeric. The value for replacing NA values
varies depending on the selected method, and each method applies different number ranges.
\enumerate{
\item QRILC: 0.1-1
\item SVD: 1-10
\item KNN: 1-10
\item min: 0.1-0.5
\item PPCA: 1-10
\item BPCA: 1-10
}
Default is \code{0.5} for replace_na_method='min'.}

\item{normalization}{Character. Normalization function. Allowed methods
include "Percentage", "PQN", "Quantile", "Sum", "Median", and "none".
If you have already normalized the abundance values, select 'none'.
Default is \code{'Percentage'}.}
}
\value{
Return a SummarizedExperiment object with processed abundance values.
}
\description{
This function processes the abundance data based on user options,
including removing features with missing values, imputing missing values, and
normalization.
}
\examples{
data("de_data_twoGroup")
processed_data <- data_process(de_data_twoGroup, exclude_missing=TRUE,
    exclude_missing_pct=70, replace_na_method='min', replace_na_method_ref=0.5,
    normalization='Percentage')
}
