% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process.R
\name{data_process}
\alias{data_process}
\title{data_process}
\usage{
data_process(
  exp_data_SE,
  exclude_var_missing = TRUE,
  missing_pct_limit = 50,
  replace_zero = TRUE,
  zero2what = "min",
  xmin = 0.5,
  replace_NA = TRUE,
  NA2what = "min",
  ymin = 0.5,
  pct_transform = TRUE,
  data_transform = TRUE,
  trans_type = "log",
  centering = FALSE,
  scaling = FALSE
)
}
\arguments{
\item{exp_data_SE}{A SummarizedExperiment object contains information about
various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
allowed. The row data  corresponds to specific lipid features, such as class
and total length. The first column's name must be "feature" (lipid species),
and NAs are allowed for this data. The column data comprises sample names,
sample labels, group names, and pair numbers that represent 'the pair' for
conducting t-tests or Wilcoxon tests. NAs are allowed.}

\item{exclude_var_missing}{Logical. Remove Lipid with too many
missing values. (default: TRUE)}

\item{missing_pct_limit}{An integer indicating the missing values over a
certain percentage should be removed. (default: 50)}

\item{replace_zero}{Logical. Replace 0. (default: TRUE)}

\item{zero2what}{A character string indicating the value to replace 0.
Value include "mean", "median", "min". (default: min)}

\item{xmin}{A numeric value indicating the min value to replace 0.}

\item{replace_NA}{Logical. If remove_na = TRUE, all NA will be removed.
(default: TRUE)}

\item{NA2what}{A character string indicating the value to replace NA.
Value include "mean", "median", "min". (default: min)}

\item{ymin}{A numeric value indicating the min value to replace NA.}

\item{pct_transform}{Logical. If \bold{pct_transform = TRUE}, transform
lipid value into a percentage. (default:TRUE)}

\item{data_transform}{Logical. If \bold{data_transform = TRUE},
transform exp_data by log10.}

\item{trans_type}{A character string of transformation type.}

\item{centering}{A logical value indicating whether the variables should be
shifted to be zero centered. Alternately, a vector of length equal to
the number of columns of x can be supplied. The value is passed to scale.
(default: FALSE)}

\item{scaling}{A logical value. If scaling = TRUE, each block is
standardized to zero means and unit variances. (default: FALSE).}
}
\value{
Return a SummarizedExperiment object.
}
\description{
This function conducts data processing according to users'
    options, including removing features with missing values, missing values
    imputation, percentage transformation, log10 transformation, etc.
}
\examples{
data("DE_data")
exp_data_SE <- DE_data
data_process(exp_data_SE, exclude_var_missing=TRUE, missing_pct_limit=50,
    replace_zero=TRUE, zero2what='min', xmin=0.5, replace_NA=TRUE,
    NA2what='min', ymin=0.5, pct_transform=TRUE, data_transform=TRUE,
    trans_type='log', centering=FALSE, scaling=FALSE)
}
