\name{DE_pathway_gene_list}
\alias{DE_pathway_gene_list}
\docType{data}
\title{
DE_pathway_gene_list
}
\description{
A list that comprises the genes included in each pathway for compiling the
significant network of the selected lipid class.
}
\usage{data("DE_pathway_gene_list")}
\value{
  The format is:
List of 307
 $ hsa:00010: chr [1:68] "ENTREZID:10327" "ENTREZID:124" "ENTREZID:125" ...
 $ hsa:00020: chr [1:30] "ENTREZID:1431" "ENTREZID:1737" "ENTREZID:1738"  ...
 $ hsa:00030: chr [1:29] "ENTREZID:2203" "ENTREZID:221823" "ENTREZID:226"  ...
 $ hsa:00040: chr [1:33] "ENTREZID:10327" "ENTREZID:10720" "ENTREZID:10941" ...
 $ hsa:00051: chr [1:32] "ENTREZID:197258" "ENTREZID:2203" "ENTREZID:226"  ...
 $ hsa:00052: chr [1:28] "ENTREZID:130589" "ENTREZID:231"  ...
 $ hsa:00053: chr [1:21] "ENTREZID:55586" "ENTREZID:7358"  ...
 $ hsa:00061: chr [1:18] "ENTREZID:7923" "ENTREZID:109703458"  ...
 $ hsa:00062: chr [1:27] "ENTREZID:10449" "ENTREZID:1892"  ...
 $ hsa:00071: chr [1:42] "ENTREZID:10449" "ENTREZID:124" ...
 $ hsa:00072: chr [1:10] "ENTREZID:3155" "ENTREZID:3157" ...
 $ hsa:00100: chr [1:20] "ENTREZID:1056" "ENTREZID:10682"  ...
 $ hsa:00120: chr [1:17] "ENTREZID:10005" "ENTREZID:10858"  ...
 $ hsa:00130: chr [1:10] "ENTREZID:10229" "ENTREZID:154807" ...
 $ hsa:00140: chr [1:61] "ENTREZID:100861540" "ENTREZID:10720"  ...
 $ hsa:00190: chr [1:47] "ENTREZID:10312" "ENTREZID:10476"  ...
 $ hsa:00220: chr [1:21] "ENTREZID:100526760" "ENTREZID:1373"  ...
 $ hsa:00230: chr [1:126] "ENTREZID:100" "ENTREZID:100526794"  ...
 $ hsa:00232: chr [1:5] "ENTREZID:10" "ENTREZID:1544" "ENTREZID:1548"  ...
 $ hsa:00240: chr [1:57] "ENTREZID:100526794" "ENTREZID:10201"  ...
 $ hsa:00250: chr [1:36] "ENTREZID:122622" "ENTREZID:1373" "ENTREZID:159" ...
 $ hsa:00260: chr [1:35] "ENTREZID:102724560" "ENTREZID:10993" ...
 $ hsa:00270: chr [1:47] "ENTREZID:102724560" "ENTREZID:1036" ...
 $ hsa:00280: chr [1:48] "ENTREZID:10449" "ENTREZID:11112" "ENTREZID:1629" ...
 $ hsa:00310: chr [1:57] "ENTREZID:10157" "ENTREZID:10919"  ...
 $ hsa:00330: chr [1:50] "ENTREZID:102724788" "ENTREZID:112483"  ...
 $ hsa:00340: chr [1:22] "ENTREZID:10841" "ENTREZID:131669" "ENTREZID:138199"...
 $ hsa:00350: chr [1:35] "ENTREZID:124" "ENTREZID:125" "ENTREZID:126" ...
 $ hsa:00360: chr [1:16] "ENTREZID:1644" "ENTREZID:218" "ENTREZID:220" ...
 $ hsa:00380: chr [1:42] "ENTREZID:121278" "ENTREZID:125061" ...
 $ hsa:00400: chr [1:5] "ENTREZID:259307" "ENTREZID:2805" "ENTREZID:2806" ...
 $ hsa:00410: chr [1:31] "ENTREZID:18" "ENTREZID:1807" "ENTREZID:1892" ...
 $ hsa:00430: chr [1:11] "ENTREZID:1036" "ENTREZID:124975" "ENTREZID:2571" ...
 $ hsa:00440: chr [1:6] "ENTREZID:5130" "ENTREZID:5833" "ENTREZID:9468" ...
 $ hsa:00450: chr [1:14] "ENTREZID:10587" "ENTREZID:114112" ...
 $ hsa:00471: chr [1:5] "ENTREZID:27165" "ENTREZID:2744" "ENTREZID:2746"  ...
 $ hsa:00480: chr [1:53] "ENTREZID:119391" "ENTREZID:124975" ...
 $ hsa:00500: chr [1:36] "ENTREZID:11181" "ENTREZID:178" "ENTREZID:2538"  ...
 $ hsa:00510: chr [1:50] "ENTREZID:10905" "ENTREZID:11253" "ENTREZID:144245" ...
 $ hsa:00512: chr [1:31] "ENTREZID:100528030" "ENTREZID:11226"  ...
 $ hsa:00513: chr [1:17] "ENTREZID:10905" "ENTREZID:11253" "ENTREZID:4121" ...
 $ hsa:00515: chr [1:23] "ENTREZID:10329" "ENTREZID:10585" "ENTREZID:11041" ...
 $ hsa:00520: chr [1:43] "ENTREZID:10007" "ENTREZID:10020" "ENTREZID:1118" ...
 $ hsa:00531: chr [1:19] "ENTREZID:10855" "ENTREZID:138050"  ...
 $ hsa:00532: chr [1:11] "ENTREZID:11285" "ENTREZID:126792" "ENTREZID:22856" ...
 $ hsa:00534: chr [1:5] "ENTREZID:2131" "ENTREZID:2132" "ENTREZID:2135" ...
 $ hsa:00561: chr [1:59] "ENTREZID:10327" "ENTREZID:10554" ...
 $ hsa:00562: chr [1:72] "ENTREZID:10423" "ENTREZID:113026"  ...
 $ hsa:00563: chr [1:23] "ENTREZID:10026" "ENTREZID:128869"  ...
 $ hsa:00564: chr [1:97] "ENTREZID:100137049" "ENTREZID:10162"  ...
 $ hsa:00565: chr [1:46] "ENTREZID:100137049" "ENTREZID:10390"  ...
 $ hsa:00590: chr [1:63] "ENTREZID:100137049" "ENTREZID:10728"  ...
 $ hsa:00591: chr [1:29] "ENTREZID:100137049" "ENTREZID:11145"  ...
 $ hsa:00592: chr [1:25] "ENTREZID:100137049" "ENTREZID:11145" ...
 $ hsa:00600: chr [1:47] "ENTREZID:10558" "ENTREZID:10715"  ...
 $ hsa:00601: chr [1:27] "ENTREZID:10317" "ENTREZID:10331"  ...
 $ hsa:00603: chr [1:14] "ENTREZID:10317" "ENTREZID:127550"  ...
 $ hsa:00604: chr [1:15] "ENTREZID:256435" "ENTREZID:2583"  ...
 $ hsa:00620: chr [1:39] "ENTREZID:10873" "ENTREZID:134526"  ...
 $ hsa:00630: chr [1:28] "ENTREZID:112817" "ENTREZID:132158" ...
 $ hsa:00640: chr [1:34] "ENTREZID:160287" "ENTREZID:1629" "ENTREZID:1738" ...
 $ hsa:00650: chr [1:27] "ENTREZID:116285" "ENTREZID:123876" ...
 $ hsa:00670: chr [1:20] "ENTREZID:100528021" "ENTREZID:10588" ...
 $ hsa:00730: chr [1:15] "ENTREZID:122481" "ENTREZID:158067" "ENTREZID:203"  ...
 $ hsa:00740: chr [1:8] "ENTREZID:5167" "ENTREZID:5169" "ENTREZID:52" ...
 $ hsa:00750: chr [1:5] "ENTREZID:316" "ENTREZID:493911" "ENTREZID:55163" ...
 $ hsa:00760: chr [1:36] "ENTREZID:100526794" "ENTREZID:10135"  ...
 $ hsa:00770: chr [1:17] "ENTREZID:1806" "ENTREZID:1807" "ENTREZID:5167"  ...
 $ hsa:00780: chr [1:2] "ENTREZID:3141" "ENTREZID:686"
 $ hsa:00785: chr [1:4] "ENTREZID:11019" "ENTREZID:116285" "ENTREZID:387787"...
 $ hsa:00790: chr [1:26] "ENTREZID:10243" "ENTREZID:121278" "ENTREZID:1719" ...
 $ hsa:00830: chr [1:67] "ENTREZID:100861540" "ENTREZID:10170"  ...
 $ hsa:00860: chr [1:39] "ENTREZID:10720" "ENTREZID:10941" "ENTREZID:1352" ...
 $ hsa:00900: chr [1:21] "ENTREZID:10269" "ENTREZID:10654" "ENTREZID:116150" ...
 $ hsa:00910: chr [1:4] "ENTREZID:1373" "ENTREZID:2746" "ENTREZID:2747" ...
 $ hsa:00920: chr [1:10] "ENTREZID:10380" "ENTREZID:23474"  ...
 $ hsa:00970: chr [1:13] "ENTREZID:118672" "ENTREZID:283459"  ...
 $ hsa:00980: chr [1:75] "ENTREZID:10720" "ENTREZID:107987478"  ...
 $ hsa:00982: chr [1:72] "ENTREZID:10720" "ENTREZID:107987478"  ...
 $ hsa:00983: chr [1:79] "ENTREZID:10" "ENTREZID:10201"  ...
 $ hsa:03008: chr [1:3] "ENTREZID:3692" "ENTREZID:51119" "ENTREZID:79631"
 $ hsa:03013: chr [1:127] "ENTREZID:10189" "ENTREZID:10284" ...
 $ hsa:03015: chr [1:70] "ENTREZID:10482" "ENTREZID:10767"  ...
 $ hsa:03018: chr [1:18] "ENTREZID:132430" "ENTREZID:140886"  ...
 $ hsa:03320: chr [1:69] "ENTREZID:5465" "ENTREZID:5467" ...
 $ hsa:03440: chr [1:18] "ENTREZID:10111" "ENTREZID:29086" ...
 $ hsa:03460: chr [1:40] "ENTREZID:100526739" "ENTREZID:116028" ...
 $ hsa:04010: chr [1:294] "ENTREZID:10000" "ENTREZID:10125"  ...
 $ hsa:04012: chr [1:85] "ENTREZID:10000" "ENTREZID:10298" ...
 $ hsa:04014: chr [1:229] "ENTREZID:10000" "ENTREZID:100271927"  ...
 $ hsa:04015: chr [1:206] "ENTREZID:10235" "ENTREZID:10411"  ...
 $ hsa:04020: chr [1:190] "ENTREZID:10105" "ENTREZID:102723407"  ...
 $ hsa:04022: chr [1:157] "ENTREZID:10000" "ENTREZID:10242" ...
 $ hsa:04024: chr [1:211] "ENTREZID:10411" "ENTREZID:10451"  ...
 $ hsa:04060: chr [1:284] "ENTREZID:10220" "ENTREZID:10344"  ...
 $ hsa:04061: chr [1:97] "ENTREZID:10344" "ENTREZID:10563" ...
 $ hsa:04062: chr [1:188] "ENTREZID:10000" "ENTREZID:10235"...
 $ hsa:04064: chr [1:92] "ENTREZID:102723407" "ENTREZID:10454" ...
 $ hsa:04066: chr [1:108] "ENTREZID:10000" "ENTREZID:112398"...
  [list output truncated]
}
\examples{
data(DE_pathway_gene_list)
}
\keyword{datasets}
