% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DE_species_2.R
\name{DE_species_2}
\alias{DE_species_2}
\title{DE_species_2}
\usage{
DE_species_2(
  exp_data,
  data_transform = TRUE,
  group_info,
  paired = FALSE,
  test = "t.test",
  adjust_p_method = "BH",
  sig_stat = "p.adj",
  sig_pvalue = 0.05,
  sig_FC = 2,
  plotly = TRUE
)
}
\arguments{
\item{exp_data}{A data frame includes the expression of lipid features in
each sample. NAs are allowed. First column should be gene/lipid name and
first column name must be 'feature'.}

\item{data_transform}{Logical. If data_transform = TRUE,
transform exp_data by log10.}

\item{group_info}{A data frame comprises the name of the sample,
the label of the sample, the group name of the sample, and the pair number
represents 'the pair' for the t-test/Wilcoxon test. NAs are allowed.}

\item{paired}{Logical. If paired = TRUE,
data are paired samples. (default: FALSE)}

\item{test}{A character string indicating which method to be used for
comparing means. Allowed method include \bold{"t.test"} and
\bold{"wilcox.test"}. (default: "t.test")}

\item{adjust_p_method}{Correction method, a character string. One of
\bold{"holm"}, \bold{"hochberg"}, \bold{"hommel"}, \bold{"bonferroni"},
\bold{"BH"}, \bold{"BY"}, \bold{"fdr"}, \bold{"none"}, can be
abbreviated. (default: "BH")}

\item{sig_stat}{A character string indicating which p-value is to be used
for the statistically significant. One of \bold{"p.adj"} or
\bold{"p"}. (default: "p.adj")}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{sig_FC}{Numeric. Significance of the fold-change. (default: 2)}

\item{plotly}{Logical value. If TRUE, return the resulting plots dynamically.}
}
\value{
Return a list with 2 data frames and 1 figures.
\enumerate{
\item DE_species_table_all: a data frame comprises the results of
differential expression analysis, including fold change, p-value,
adjusted p-value.
\item DE_species_table_sig: a data frame comprises the significant
results of differential expression analysis, including fold change,
p-value, adjusted p-value.
\item DE_species_dotchart_sig: a lollipop chart reveals the lipid species
that pass chosen cut-offs.
\item DE_species_maplot: an MA plot of lipid species
\item DE_species_volcano: a volcano plot of lipid species
}
}
\description{
Compute differentially expressed analysis of
two groups (independent) based on the input Group Information to find
significant lipid species. To find differentially expressed lipid species,
a statistical method (t-test or Wilcoxon test) and the cut-offs for
significant lipid species need to be chosen, and the p-value will then be
adjusted by Benjamini-Hochberg or other methods.
}
\examples{
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
exp_data <- DE_exp_data
lipid_char_table <- DE_lipid_char_table
group_info <- DE_group_info
exp_transform_non_log <- data_process(exp_data, exclude_var_missing=TRUE,
                                      missing_pct_limit=50,
                                      replace_zero=TRUE, zero2what='min',
                                      xmin=0.5, replace_NA=TRUE,
                                      NA2what='min', ymin=0.5,
                                      pct_transform=TRUE,
                                      data_transform=FALSE,
                                      trans_type='log',
                                      centering=FALSE, scaling=FALSE)
DE_species_2(exp_transform_non_log, data_transform=TRUE,
             group_info=group_info, paired=FALSE,
             test='t.test', adjust_p_method='BH', sig_stat='p.adj',
             sig_pvalue=0.05, sig_FC=2, plotly=TRUE)
}
