% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Enrichment}
\usage{
Enrichment(
  DE_species_table_sig,
  lipid_char_table,
  char_var,
  sig_pvalue = 0.05,
  plotly = TRUE
)
}
\arguments{
\item{DE_species_table_sig}{A data frame comprises the significant results
of differential expression analysis, including fold change, p-value,
adjusted p-value. The output of \code{\link{DE_species_2}}.}

\item{lipid_char_table}{A data frame with lipid features, such as class,
total length. NAs are allowed. The name of first column must be "feature".}

\item{char_var}{A character string of the first lipid characteristic
selected by users from the column name of \bold{lipid_char_table},
such as total length.}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}

\item{plotly}{Logical value. If TRUE, return the resulting plots dynamically.}
}
\value{
Return a list with 1 tibble and 1 plot.
\enumerate{
\item enrich_char_table: a data frame encompasses condition, characteristic,
 p-value, and significance.
\item enrich_char_barplot: the enrichment plot for
up/down/non-significant groups
}
}
\description{
\enumerate{
\item Determine whether significant lipid species are enriched in
user-selected characteristics.
\item Print a summary table and enrichment bar chart.
}
}
\examples{
library(magrittr)
library(dplyr)
data("DE_exp_data")
data("DE_lipid_char_table")
data("DE_group_info")
exp_data <- DE_exp_data
lipid_char_table <- DE_lipid_char_table
group_info <- DE_group_info
exp_transform <- data_process(exp_data, exclude_var_missing=TRUE,
                              missing_pct_limit=50, replace_zero=TRUE,
                              zero2what='min', xmin=0.5, replace_NA=TRUE,
                              NA2what='min', ymin=0.5, pct_transform=TRUE,
                              data_transform=TRUE, trans_type='log',
                              centering=FALSE,  scaling=FALSE)
exp_transform_non_log <- data_process(exp_data, exclude_var_missing=TRUE,
                                      missing_pct_limit=50,
                                      replace_zero=TRUE, zero2what='min',
                                      xmin=0.5, replace_NA=TRUE,
                                      NA2what='min', ymin=0.5,
                                      pct_transform=TRUE,
                                      data_transform=FALSE,
                                      trans_type='log', centering=FALSE,
                                      scaling=FALSE)
lipid_char_filter <- lipid_char_table \%>\%
    filter(feature \%in\% exp_transform$feature)
DE_species_table_sig <- DE_species_2(exp_transform_non_log,
                                     data_transform=TRUE,
                                     group_info=group_info, paired=FALSE,
                                     test='t.test', adjust_p_method='BH',
                                     sig_stat='p.adj', sig_pvalue=0.05,
                                     sig_FC=2,
                                     plotly=TRUE)$DE_species_table_sig
char_var <- colnames(lipid_char_filter)[-1]
Enrichment (DE_species_table_sig, lipid_char_table=lipid_char_filter,
            char_var=char_var[1], sig_pvalue=0.05, plotly=TRUE)
}
