% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\name{Enrichment}
\alias{Enrichment}
\title{Enrichment}
\usage{
Enrichment(DE_species_sig, exp_transform_SE, char_var, sig_pvalue = 0.05)
}
\arguments{
\item{DE_species_sig}{A SummarizedExperiment object comprises the significant
results of differential expression analysis, including fold change, p-value,
adjusted p-value. The output of \code{\link{DE_species}}.}

\item{exp_transform_SE}{A SummarizedExperiment object contains information
about various features, such as molecules, lipid class, etc., and their
expression levels for each sample. The assay is a matrix representing the
expression of lipid features across all samples. Missing values (NAs)  are
not allowed. The row data corresponds to specific lipid features, such as
class and total length. The first column's name must be "feature" (lipid
species), and NAs are allowed for this data. The column data comprises
sample names, sample labels, group names, and pair numbers that represent
'the pair' for conducting t-tests or Wilcoxon tests. NAs are allowed.}

\item{char_var}{A character string of the first lipid characteristic
selected by users from the column name of the rowData of \bold{exp_data_SE},
such as total length.}

\item{sig_pvalue}{Numeric. Significant level. (default: 0.05)}
}
\value{
Return a SummarizedExperiment object.
}
\description{
Determine whether significant lipid species are enriched in
user-selected characteristics.
}
\examples{
library(dplyr)
library(SummarizedExperiment)
data("DE_data")
exp_data_SE <- DE_data
exp_transform_SE <- data_process(exp_data_SE, exclude_var_missing=TRUE,
    missing_pct_limit=50, replace_zero=TRUE, zero2what='min', xmin=0.5,
    replace_NA=TRUE, NA2what='min', ymin=0.5, pct_transform=TRUE,
    data_transform=TRUE, trans_type='log', centering=FALSE, scaling=FALSE)
exp_transform_non_log <- data_process(exp_data_SE, exclude_var_missing=TRUE,
    missing_pct_limit=50, replace_zero=TRUE, zero2what='min', xmin=0.5,
    replace_NA=TRUE, NA2what='min', ymin=0.5, pct_transform=TRUE,
    data_transform=FALSE, trans_type='log', centering=FALSE, scaling=FALSE)
DE_species_result <- DE_species(exp_transform_non_log, data_transform=TRUE,
    paired=FALSE, test='t.test', adjust_p_method='BH', sig_stat='p.adj',
    sig_pvalue=0.05, sig_FC=2)
lipid_char_table <- as.data.frame(rowData(exp_transform_SE))
char_var <- colnames(lipid_char_table)[-1]
DE_result_sig <- DE_species_result$exp_data_stat_sig
Enrichment(DE_result_sig, exp_transform_SE, char_var=char_var[1],
    sig_pvalue=0.05)
}
