\name{profiling_exp_data}
\alias{profiling_exp_data}
\docType{data}
\title{
profiling_exp_data
}
\description{
A data frame of lipid expression data for profiling analysis workflow. The lipid
expression data were obtained from the research of adipose tissue ATGL modifies
the cardiac lipidome in pressure-overload-induced left Ventricular failure.
Human plasma lipidome from 10 healthy controls and 13 patients with systolic
heart failure (HFrEF) were analyzed by MS-based shotgun lipidomics.
}
\usage{data("profiling_exp_data")}
\value{
  A data frame with 208 observations on the following 24 variables.
  \describe{
    \item{\code{feature}}{a character vector}
    \item{\code{control_01}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_02}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_03}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_04}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_05}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_06}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_07}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_08}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_09}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{control_10}}{a numeric vector,
    expression value of one of the 10 healthy controls}
    \item{\code{hfref_patient_01}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_02}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_03}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_04}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_05}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_06}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_07}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_08}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_09}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_10}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_11}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_12}}{a numeric vector,
    expression value of one of the 13 patients}
    \item{\code{hfref_patient_13}}{a numeric vector,
    expression value of one of the 13 patients}
  }
}
\source{
Salatzki J, Foryst-Ludwig A, Bentele K, Blumrich A, Smeir E, Ban Z, et al.
Adipose tissue ATGL modifies the cardiac lipidome in pressure-overload-induced
left ventricular failure. PLoS genetics. 2018;14(1):e1007171.
}
\examples{
data(profiling_exp_data)
}
\keyword{datasets}
