% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SHAP_dependence_plot.R
\name{SHAP_dependence_plot}
\alias{SHAP_dependence_plot}
\title{SHAP_dependence_plot}
\usage{
SHAP_dependence_plot(shap_long, x, y, color_var)
}
\arguments{
\item{shap_long}{The output data frame of \code{\link{SHAP}} (list[[2]]).}

\item{x}{A character string of the feature name selected by users from the "variable" column of \bold{shap_long}.}

\item{y}{A character string of the feature name selected by users from the "variable" column of \bold{shap_long}.}

\item{color_var}{A character string of the feature name selected by users from the "variable" column of \bold{shap_long}.}
}
\value{
Return a list of 1 plot.
\enumerate{
\item shap_dependence_plot: SHAP dependence plot.
}
}
\description{
This function visualizes the SHAP values against the feature values for each variable.
}
\examples{
library(SHAPforxgboost)
data("ML_exp_data")
data("ML_lipid_char_table")
data("ML_condition_table")
exp_data <- ML_exp_data
lipid_char_table <- ML_lipid_char_table
condition_table <- ML_condition_table
char_var <- colnames(lipid_char_table)[-1]
ML_data <- ML_data_process(exp_data, group_info = condition_table,
                           lipid_char_table, char_var[1],
                           exclude_var_missing=TRUE, missing_pct_limit=50,
                           replace_zero=TRUE, zero2what='min', xmin=0.5,
                           replace_NA=TRUE, NA2what='min', ymin=0.5,
                           pct_transform=TRUE, data_transform=TRUE,
                           trans_type='log', centering=FALSE, scaling=FALSE)
ML_output <- ML_final(ML_data[[2]], ranking_method='Random_forest',
                      ML_method='Random_forest', split_prop=0.3, nfold=10)
SHAP_output <- SHAP(ML_data[[2]], best_model=ML_output[[8]],
                    best_model_feature=ML_output[[9]],
                    ML_method='Random_forest', feature_n=10, nsim=5)
SHAP_dependence_plot(SHAP_output[[2]],x="C38.6.PC", y="C38.6.PC",
                     color_var="C38.6.PC")
}
